Struct aws_sdk_redshift::client::fluent_builders::RestoreFromClusterSnapshot

pub struct RestoreFromClusterSnapshot<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> RestoreFromClusterSnapshot<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<RestoreFromClusterSnapshotOutput, SdkError<RestoreFromClusterSnapshotError>> where
R::Policy: SmithyRetryPolicy<RestoreFromClusterSnapshotInputOperationOutputAlias, RestoreFromClusterSnapshotOutput, RestoreFromClusterSnapshotError, RestoreFromClusterSnapshotInputOperationRetryAlias>,

pub fn cluster_identifier(self, inp: impl Into<String>) -> Self

The identifier of the cluster that will be created from restoring the snapshot.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

pub fn set_cluster_identifier(self, input: Option<String>) -> Self

pub fn snapshot_identifier(self, inp: impl Into<String>) -> Self

The name of the snapshot from which to create the new cluster. This parameter isn't case sensitive.

Example: my-snapshot-id

pub fn set_snapshot_identifier(self, input: Option<String>) -> Self

pub fn snapshot_cluster_identifier(self, inp: impl Into<String>) -> Self

The name of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

pub fn set_snapshot_cluster_identifier(self, input: Option<String>) -> Self

pub fn port(self, inp: i32) -> Self

The port number on which the cluster accepts connections.

Default: The same port as the original cluster.

Constraints: Must be between 1115 and 65535.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn availability_zone(self, inp: impl Into<String>) -> Self

The Amazon EC2 Availability Zone in which to restore the cluster.

Default: A random, system-chosen Availability Zone.

Example: us-east-2a

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn allow_version_upgrade(self, inp: bool) -> Self

If true, major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster.

Default: true

pub fn set_allow_version_upgrade(self, input: Option<bool>) -> Self

pub fn cluster_subnet_group_name(self, inp: impl Into<String>) -> Self

The name of the subnet group where you want to cluster restored.

A snapshot of cluster in VPC can be restored only in VPC. Therefore, you must provide subnet group name where you want the cluster restored.

pub fn set_cluster_subnet_group_name(self, input: Option<String>) -> Self

pub fn publicly_accessible(self, inp: bool) -> Self

If true, the cluster can be accessed from a public network.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn owner_account(self, inp: impl Into<String>) -> Self

The Amazon Web Services account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

pub fn set_owner_account(self, input: Option<String>) -> Self

pub fn hsm_client_certificate_identifier(self, inp: impl Into<String>) -> Self

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

pub fn set_hsm_client_certificate_identifier(
self,
input: Option<String>
) -> Self

pub fn hsm_configuration_identifier(self, inp: impl Into<String>) -> Self

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

pub fn set_hsm_configuration_identifier(self, input: Option<String>) -> Self

pub fn elastic_ip(self, inp: impl Into<String>) -> Self

The elastic IP (EIP) address for the cluster.

pub fn set_elastic_ip(self, input: Option<String>) -> Self

pub fn cluster_parameter_group_name(self, inp: impl Into<String>) -> Self

The name of the parameter group to be associated with this cluster.

Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

pub fn set_cluster_parameter_group_name(self, input: Option<String>) -> Self

pub fn cluster_security_groups(self, inp: impl Into<String>) -> Self

Appends an item to ClusterSecurityGroups.

To override the contents of this collection use set_cluster_security_groups.

A list of security groups to be associated with this cluster.

Default: The default cluster security group for Amazon Redshift.

Cluster security groups only apply to clusters outside of VPCs.

pub fn set_cluster_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to VpcSecurityGroupIds.

To override the contents of this collection use set_vpc_security_group_ids.

A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster.

Default: The default VPC security group is associated with the cluster.

VPC security groups only apply to clusters in VPCs.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn preferred_maintenance_window(self, inp: impl Into<String>) -> Self

The weekly time range (in UTC) during which automated cluster maintenance can occur.

Format: ddd:hh24:mi-ddd:hh24:mi

Default: The value selected for the cluster from which the snapshot was taken. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Minimum 30-minute window.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn automated_snapshot_retention_period(self, inp: i32) -> Self

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: The value selected for the cluster from which the snapshot was taken.

Constraints: Must be a value from 0 to 35.

pub fn set_automated_snapshot_retention_period(self, input: Option<i32>) -> Self

pub fn manual_snapshot_retention_period(self, inp: i32) -> Self

The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

pub fn set_manual_snapshot_retention_period(self, input: Option<i32>) -> Self

pub fn kms_key_id(self, inp: impl Into<String>) -> Self

The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster that you restore from a shared snapshot.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn node_type(self, inp: impl Into<String>) -> Self

The node type that the restored cluster will be provisioned with.

Default: The node type of the cluster from which the snapshot was taken. You can modify this if you are using any DS node type. In that case, you can choose to restore into another DS node type of the same size. For example, you can restore ds1.8xlarge into ds2.8xlarge, or ds1.xlarge into ds2.xlarge. If you have a DC instance type, you must restore into that same instance type and size. In other words, you can only restore a dc1.large instance type into another dc1.large instance type or dc2.large instance type. You can't restore dc1.8xlarge to dc2.8xlarge. First restore to a dc1.8xlarge cluster, then resize to a dc2.8large cluster. For more information about node types, see About Clusters and Nodes in the Amazon Redshift Cluster Management Guide.

pub fn set_node_type(self, input: Option<String>) -> Self

pub fn enhanced_vpc_routing(self, inp: bool) -> Self

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

pub fn set_enhanced_vpc_routing(self, input: Option<bool>) -> Self

pub fn additional_info(self, inp: impl Into<String>) -> Self

Reserved.

pub fn set_additional_info(self, input: Option<String>) -> Self

pub fn iam_roles(self, inp: impl Into<String>) -> Self

Appends an item to IamRoles.

To override the contents of this collection use set_iam_roles.

A list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request.

A cluster can have up to 10 IAM roles associated at any time.

pub fn set_iam_roles(self, input: Option<Vec<String>>) -> Self

pub fn maintenance_track_name(self, inp: impl Into<String>) -> Self

The name of the maintenance track for the restored cluster. When you take a snapshot, the snapshot inherits the MaintenanceTrack value from the cluster. The snapshot might be on a different track than the cluster that was the source for the snapshot. For example, suppose that you take a snapshot of a cluster that is on the current track and then change the cluster to be on the trailing track. In this case, the snapshot and the source cluster are on different tracks.

pub fn set_maintenance_track_name(self, input: Option<String>) -> Self

pub fn snapshot_schedule_identifier(self, inp: impl Into<String>) -> Self

A unique identifier for the snapshot schedule.

pub fn set_snapshot_schedule_identifier(self, input: Option<String>) -> Self

pub fn number_of_nodes(self, inp: i32) -> Self

The number of nodes specified when provisioning the restored cluster.

pub fn set_number_of_nodes(self, input: Option<i32>) -> Self

pub fn availability_zone_relocation(self, inp: bool) -> Self

The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster is restored.

pub fn set_availability_zone_relocation(self, input: Option<bool>) -> Self

pub fn aqua_configuration_status(self, inp: AquaConfigurationStatus) -> Self

The value represents how the cluster is configured to use AQUA (Advanced Query Accelerator) after the cluster is restored. Possible values include the following.

enabled - Use AQUA if it is available for the current Amazon Web Services Region and Amazon Redshift node type.

disabled - Don't use AQUA.

auto - Amazon Redshift determines whether to use AQUA.

pub fn set_aqua_configuration_status(
self,
input: Option<AquaConfigurationStatus>
) -> Self