Struct aws_sdk_redshift::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for RedshiftServiceVersion20121201.

This client allows ergonomic access to a RedshiftServiceVersion20121201-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_reserved_node_exchange(
&self
) -> AcceptReservedNodeExchange<C, M, R>

pub fn add_partner(&self) -> AddPartner<C, M, R>

pub fn associate_data_share_consumer(
&self
) -> AssociateDataShareConsumer<C, M, R>

pub fn authorize_cluster_security_group_ingress(
&self
) -> AuthorizeClusterSecurityGroupIngress<C, M, R>

pub fn authorize_data_share(&self) -> AuthorizeDataShare<C, M, R>

pub fn authorize_endpoint_access(&self) -> AuthorizeEndpointAccess<C, M, R>

pub fn authorize_snapshot_access(&self) -> AuthorizeSnapshotAccess<C, M, R>

pub fn batch_delete_cluster_snapshots(
&self
) -> BatchDeleteClusterSnapshots<C, M, R>

pub fn batch_modify_cluster_snapshots(
&self
) -> BatchModifyClusterSnapshots<C, M, R>

pub fn cancel_resize(&self) -> CancelResize<C, M, R>

pub fn copy_cluster_snapshot(&self) -> CopyClusterSnapshot<C, M, R>

pub fn create_authentication_profile(
&self
) -> CreateAuthenticationProfile<C, M, R>

pub fn create_cluster(&self) -> CreateCluster<C, M, R>

pub fn create_cluster_parameter_group(
&self
) -> CreateClusterParameterGroup<C, M, R>

pub fn create_cluster_security_group(
&self
) -> CreateClusterSecurityGroup<C, M, R>

pub fn create_cluster_snapshot(&self) -> CreateClusterSnapshot<C, M, R>

pub fn create_cluster_subnet_group(&self) -> CreateClusterSubnetGroup<C, M, R>

pub fn create_endpoint_access(&self) -> CreateEndpointAccess<C, M, R>

pub fn create_event_subscription(&self) -> CreateEventSubscription<C, M, R>

pub fn create_hsm_client_certificate(
&self
) -> CreateHsmClientCertificate<C, M, R>

pub fn create_hsm_configuration(&self) -> CreateHsmConfiguration<C, M, R>

pub fn create_scheduled_action(&self) -> CreateScheduledAction<C, M, R>

pub fn create_snapshot_copy_grant(&self) -> CreateSnapshotCopyGrant<C, M, R>

pub fn create_snapshot_schedule(&self) -> CreateSnapshotSchedule<C, M, R>

pub fn create_tags(&self) -> CreateTags<C, M, R>

pub fn create_usage_limit(&self) -> CreateUsageLimit<C, M, R>

pub fn deauthorize_data_share(&self) -> DeauthorizeDataShare<C, M, R>

pub fn delete_authentication_profile(
&self
) -> DeleteAuthenticationProfile<C, M, R>

pub fn delete_cluster(&self) -> DeleteCluster<C, M, R>

pub fn delete_cluster_parameter_group(
&self
) -> DeleteClusterParameterGroup<C, M, R>

pub fn delete_cluster_security_group(
&self
) -> DeleteClusterSecurityGroup<C, M, R>

pub fn delete_cluster_snapshot(&self) -> DeleteClusterSnapshot<C, M, R>

pub fn delete_cluster_subnet_group(&self) -> DeleteClusterSubnetGroup<C, M, R>

pub fn delete_endpoint_access(&self) -> DeleteEndpointAccess<C, M, R>

pub fn delete_event_subscription(&self) -> DeleteEventSubscription<C, M, R>

pub fn delete_hsm_client_certificate(
&self
) -> DeleteHsmClientCertificate<C, M, R>

pub fn delete_hsm_configuration(&self) -> DeleteHsmConfiguration<C, M, R>

pub fn delete_partner(&self) -> DeletePartner<C, M, R>

pub fn delete_scheduled_action(&self) -> DeleteScheduledAction<C, M, R>

pub fn delete_snapshot_copy_grant(&self) -> DeleteSnapshotCopyGrant<C, M, R>

pub fn delete_snapshot_schedule(&self) -> DeleteSnapshotSchedule<C, M, R>

pub fn delete_tags(&self) -> DeleteTags<C, M, R>

pub fn delete_usage_limit(&self) -> DeleteUsageLimit<C, M, R>

pub fn describe_account_attributes(&self) -> DescribeAccountAttributes<C, M, R>

pub fn describe_authentication_profiles(
&self
) -> DescribeAuthenticationProfiles<C, M, R>

pub fn describe_cluster_db_revisions(
&self
) -> DescribeClusterDbRevisions<C, M, R>

pub fn describe_cluster_parameter_groups(
&self
) -> DescribeClusterParameterGroups<C, M, R>

pub fn describe_cluster_parameters(&self) -> DescribeClusterParameters<C, M, R>

pub fn describe_clusters(&self) -> DescribeClusters<C, M, R>

pub fn describe_cluster_security_groups(
&self
) -> DescribeClusterSecurityGroups<C, M, R>

pub fn describe_cluster_snapshots(&self) -> DescribeClusterSnapshots<C, M, R>

pub fn describe_cluster_subnet_groups(
&self
) -> DescribeClusterSubnetGroups<C, M, R>

pub fn describe_cluster_tracks(&self) -> DescribeClusterTracks<C, M, R>

pub fn describe_cluster_versions(&self) -> DescribeClusterVersions<C, M, R>

pub fn describe_data_shares(&self) -> DescribeDataShares<C, M, R>

pub fn describe_data_shares_for_consumer(
&self
) -> DescribeDataSharesForConsumer<C, M, R>

pub fn describe_data_shares_for_producer(
&self
) -> DescribeDataSharesForProducer<C, M, R>

pub fn describe_default_cluster_parameters(
&self
) -> DescribeDefaultClusterParameters<C, M, R>

pub fn describe_endpoint_access(&self) -> DescribeEndpointAccess<C, M, R>

pub fn describe_endpoint_authorization(
&self
) -> DescribeEndpointAuthorization<C, M, R>

pub fn describe_event_categories(&self) -> DescribeEventCategories<C, M, R>

pub fn describe_events(&self) -> DescribeEvents<C, M, R>

pub fn describe_event_subscriptions(
&self
) -> DescribeEventSubscriptions<C, M, R>

pub fn describe_hsm_client_certificates(
&self
) -> DescribeHsmClientCertificates<C, M, R>

pub fn describe_hsm_configurations(&self) -> DescribeHsmConfigurations<C, M, R>

pub fn describe_logging_status(&self) -> DescribeLoggingStatus<C, M, R>

pub fn describe_node_configuration_options(
&self
) -> DescribeNodeConfigurationOptions<C, M, R>

pub fn describe_orderable_cluster_options(
&self
) -> DescribeOrderableClusterOptions<C, M, R>

pub fn describe_partners(&self) -> DescribePartners<C, M, R>

pub fn describe_reserved_node_offerings(
&self
) -> DescribeReservedNodeOfferings<C, M, R>

pub fn describe_reserved_nodes(&self) -> DescribeReservedNodes<C, M, R>

pub fn describe_resize(&self) -> DescribeResize<C, M, R>

pub fn describe_scheduled_actions(&self) -> DescribeScheduledActions<C, M, R>

pub fn describe_snapshot_copy_grants(
&self
) -> DescribeSnapshotCopyGrants<C, M, R>

pub fn describe_snapshot_schedules(&self) -> DescribeSnapshotSchedules<C, M, R>

pub fn describe_storage(&self) -> DescribeStorage<C, M, R>

pub fn describe_table_restore_status(
&self
) -> DescribeTableRestoreStatus<C, M, R>

pub fn describe_tags(&self) -> DescribeTags<C, M, R>

pub fn describe_usage_limits(&self) -> DescribeUsageLimits<C, M, R>

pub fn disable_logging(&self) -> DisableLogging<C, M, R>

pub fn disable_snapshot_copy(&self) -> DisableSnapshotCopy<C, M, R>

pub fn disassociate_data_share_consumer(
&self
) -> DisassociateDataShareConsumer<C, M, R>

pub fn enable_logging(&self) -> EnableLogging<C, M, R>

pub fn enable_snapshot_copy(&self) -> EnableSnapshotCopy<C, M, R>

pub fn get_cluster_credentials(&self) -> GetClusterCredentials<C, M, R>

pub fn get_reserved_node_exchange_offerings(
&self
) -> GetReservedNodeExchangeOfferings<C, M, R>

pub fn modify_aqua_configuration(&self) -> ModifyAquaConfiguration<C, M, R>

pub fn modify_authentication_profile(
&self
) -> ModifyAuthenticationProfile<C, M, R>

pub fn modify_cluster(&self) -> ModifyCluster<C, M, R>

pub fn modify_cluster_db_revision(&self) -> ModifyClusterDbRevision<C, M, R>

pub fn modify_cluster_iam_roles(&self) -> ModifyClusterIamRoles<C, M, R>

pub fn modify_cluster_maintenance(&self) -> ModifyClusterMaintenance<C, M, R>

pub fn modify_cluster_parameter_group(
&self
) -> ModifyClusterParameterGroup<C, M, R>

pub fn modify_cluster_snapshot(&self) -> ModifyClusterSnapshot<C, M, R>

pub fn modify_cluster_snapshot_schedule(
&self
) -> ModifyClusterSnapshotSchedule<C, M, R>

pub fn modify_cluster_subnet_group(&self) -> ModifyClusterSubnetGroup<C, M, R>

pub fn modify_endpoint_access(&self) -> ModifyEndpointAccess<C, M, R>

pub fn modify_event_subscription(&self) -> ModifyEventSubscription<C, M, R>

pub fn modify_scheduled_action(&self) -> ModifyScheduledAction<C, M, R>

pub fn modify_snapshot_copy_retention_period(
&self
) -> ModifySnapshotCopyRetentionPeriod<C, M, R>

pub fn modify_snapshot_schedule(&self) -> ModifySnapshotSchedule<C, M, R>

pub fn modify_usage_limit(&self) -> ModifyUsageLimit<C, M, R>

pub fn pause_cluster(&self) -> PauseCluster<C, M, R>

pub fn purchase_reserved_node_offering(
&self
) -> PurchaseReservedNodeOffering<C, M, R>

pub fn reboot_cluster(&self) -> RebootCluster<C, M, R>

pub fn reject_data_share(&self) -> RejectDataShare<C, M, R>

pub fn reset_cluster_parameter_group(
&self
) -> ResetClusterParameterGroup<C, M, R>

pub fn resize_cluster(&self) -> ResizeCluster<C, M, R>

pub fn restore_from_cluster_snapshot(
&self
) -> RestoreFromClusterSnapshot<C, M, R>

pub fn restore_table_from_cluster_snapshot(
&self
) -> RestoreTableFromClusterSnapshot<C, M, R>

pub fn resume_cluster(&self) -> ResumeCluster<C, M, R>

pub fn revoke_cluster_security_group_ingress(
&self
) -> RevokeClusterSecurityGroupIngress<C, M, R>

pub fn revoke_endpoint_access(&self) -> RevokeEndpointAccess<C, M, R>

pub fn revoke_snapshot_access(&self) -> RevokeSnapshotAccess<C, M, R>

pub fn rotate_encryption_key(&self) -> RotateEncryptionKey<C, M, R>

pub fn update_partner_status(&self) -> UpdatePartnerStatus<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self