Module aws_sdk_redshift::error

Modules

access_to_cluster_denied_fault

See AccessToClusterDeniedFault

access_to_snapshot_denied_fault

See AccessToSnapshotDeniedFault

authentication_profile_already_exists_fault

See AuthenticationProfileAlreadyExistsFault

authentication_profile_not_found_fault

See AuthenticationProfileNotFoundFault

authentication_profile_quota_exceeded_fault

See AuthenticationProfileQuotaExceededFault

authorization_already_exists_fault

See AuthorizationAlreadyExistsFault

authorization_not_found_fault

See AuthorizationNotFoundFault

authorization_quota_exceeded_fault

See AuthorizationQuotaExceededFault

batch_delete_request_size_exceeded_fault

See BatchDeleteRequestSizeExceededFault

batch_modify_cluster_snapshots_limit_exceeded_fault

See BatchModifyClusterSnapshotsLimitExceededFault

bucket_not_found_fault

See BucketNotFoundFault

cluster_already_exists_fault

See ClusterAlreadyExistsFault

cluster_not_found_fault

See ClusterNotFoundFault

cluster_on_latest_revision_fault

See ClusterOnLatestRevisionFault

cluster_parameter_group_already_exists_fault

See ClusterParameterGroupAlreadyExistsFault

cluster_parameter_group_not_found_fault

See ClusterParameterGroupNotFoundFault

cluster_parameter_group_quota_exceeded_fault

See ClusterParameterGroupQuotaExceededFault

cluster_quota_exceeded_fault

See ClusterQuotaExceededFault

cluster_security_group_already_exists_fault

See ClusterSecurityGroupAlreadyExistsFault

cluster_security_group_not_found_fault

See ClusterSecurityGroupNotFoundFault

cluster_security_group_quota_exceeded_fault

See ClusterSecurityGroupQuotaExceededFault

cluster_snapshot_already_exists_fault

See ClusterSnapshotAlreadyExistsFault

cluster_snapshot_not_found_fault

See ClusterSnapshotNotFoundFault

cluster_snapshot_quota_exceeded_fault

See ClusterSnapshotQuotaExceededFault

cluster_subnet_group_already_exists_fault

See ClusterSubnetGroupAlreadyExistsFault

cluster_subnet_group_not_found_fault

See ClusterSubnetGroupNotFoundFault

cluster_subnet_group_quota_exceeded_fault

See ClusterSubnetGroupQuotaExceededFault

cluster_subnet_quota_exceeded_fault

See ClusterSubnetQuotaExceededFault

copy_to_region_disabled_fault

See CopyToRegionDisabledFault

dependent_service_request_throttling_fault

See DependentServiceRequestThrottlingFault

dependent_service_unavailable_fault

See DependentServiceUnavailableFault

endpoint_already_exists_fault

See EndpointAlreadyExistsFault

endpoint_authorization_already_exists_fault

See EndpointAuthorizationAlreadyExistsFault

endpoint_authorization_not_found_fault

See EndpointAuthorizationNotFoundFault

endpoint_authorizations_per_cluster_limit_exceeded_fault

See EndpointAuthorizationsPerClusterLimitExceededFault

endpoint_not_found_fault

See EndpointNotFoundFault

endpoints_per_authorization_limit_exceeded_fault

See EndpointsPerAuthorizationLimitExceededFault

endpoints_per_cluster_limit_exceeded_fault

See EndpointsPerClusterLimitExceededFault

event_subscription_quota_exceeded_fault

See EventSubscriptionQuotaExceededFault

hsm_client_certificate_already_exists_fault

See HsmClientCertificateAlreadyExistsFault

hsm_client_certificate_not_found_fault

See HsmClientCertificateNotFoundFault

hsm_client_certificate_quota_exceeded_fault

See HsmClientCertificateQuotaExceededFault

hsm_configuration_already_exists_fault

See HsmConfigurationAlreadyExistsFault

hsm_configuration_not_found_fault

See HsmConfigurationNotFoundFault

hsm_configuration_quota_exceeded_fault

See HsmConfigurationQuotaExceededFault

in_progress_table_restore_quota_exceeded_fault

See InProgressTableRestoreQuotaExceededFault

incompatible_orderable_options

See IncompatibleOrderableOptions

insufficient_cluster_capacity_fault

See InsufficientClusterCapacityFault

insufficient_s3_bucket_policy_fault

See InsufficientS3BucketPolicyFault

invalid_authentication_profile_request_fault

See InvalidAuthenticationProfileRequestFault

invalid_authorization_state_fault

See InvalidAuthorizationStateFault

invalid_cluster_parameter_group_state_fault

See InvalidClusterParameterGroupStateFault

invalid_cluster_security_group_state_fault

See InvalidClusterSecurityGroupStateFault

invalid_cluster_snapshot_schedule_state_fault

See InvalidClusterSnapshotScheduleStateFault

invalid_cluster_snapshot_state_fault

See InvalidClusterSnapshotStateFault

invalid_cluster_state_fault

See InvalidClusterStateFault

invalid_cluster_subnet_group_state_fault

See InvalidClusterSubnetGroupStateFault

invalid_cluster_subnet_state_fault

See InvalidClusterSubnetStateFault

invalid_cluster_track_fault

See InvalidClusterTrackFault

invalid_data_share_fault

See InvalidDataShareFault

invalid_elastic_ip_fault

See InvalidElasticIpFault

invalid_endpoint_state_fault

See InvalidEndpointStateFault

invalid_hsm_client_certificate_state_fault

See InvalidHsmClientCertificateStateFault

invalid_hsm_configuration_state_fault

See InvalidHsmConfigurationStateFault

invalid_namespace_fault

See InvalidNamespaceFault

invalid_reserved_node_state_fault

See InvalidReservedNodeStateFault

invalid_restore_fault

See InvalidRestoreFault

invalid_retention_period_fault

See InvalidRetentionPeriodFault

invalid_s3_bucket_name_fault

See InvalidS3BucketNameFault

invalid_s3_key_prefix_fault

See InvalidS3KeyPrefixFault

invalid_schedule_fault

See InvalidScheduleFault

invalid_scheduled_action_fault

See InvalidScheduledActionFault

invalid_snapshot_copy_grant_state_fault

See InvalidSnapshotCopyGrantStateFault

invalid_subnet

See InvalidSubnet

invalid_subscription_state_fault

See InvalidSubscriptionStateFault

invalid_table_restore_argument_fault

See InvalidTableRestoreArgumentFault

invalid_tag_fault

See InvalidTagFault

invalid_usage_limit_fault

See InvalidUsageLimitFault

invalid_vpc_network_state_fault

See InvalidVpcNetworkStateFault

limit_exceeded_fault

See LimitExceededFault

number_of_nodes_per_cluster_limit_exceeded_fault

See NumberOfNodesPerClusterLimitExceededFault

number_of_nodes_quota_exceeded_fault

See NumberOfNodesQuotaExceededFault

partner_not_found_fault

See PartnerNotFoundFault

reserved_node_already_exists_fault

See ReservedNodeAlreadyExistsFault

reserved_node_already_migrated_fault

See ReservedNodeAlreadyMigratedFault

reserved_node_not_found_fault

See ReservedNodeNotFoundFault

reserved_node_offering_not_found_fault

See ReservedNodeOfferingNotFoundFault

reserved_node_quota_exceeded_fault

See ReservedNodeQuotaExceededFault

resize_not_found_fault

See ResizeNotFoundFault

resource_not_found_fault

See ResourceNotFoundFault

schedule_definition_type_unsupported_fault

See ScheduleDefinitionTypeUnsupportedFault

scheduled_action_already_exists_fault

See ScheduledActionAlreadyExistsFault

scheduled_action_not_found_fault

See ScheduledActionNotFoundFault

scheduled_action_quota_exceeded_fault

See ScheduledActionQuotaExceededFault

scheduled_action_type_unsupported_fault

See ScheduledActionTypeUnsupportedFault

snapshot_copy_already_disabled_fault

See SnapshotCopyAlreadyDisabledFault

snapshot_copy_already_enabled_fault

See SnapshotCopyAlreadyEnabledFault

snapshot_copy_disabled_fault

See SnapshotCopyDisabledFault

snapshot_copy_grant_already_exists_fault

See SnapshotCopyGrantAlreadyExistsFault

snapshot_copy_grant_not_found_fault

See SnapshotCopyGrantNotFoundFault

snapshot_copy_grant_quota_exceeded_fault

See SnapshotCopyGrantQuotaExceededFault

snapshot_schedule_already_exists_fault

See SnapshotScheduleAlreadyExistsFault

snapshot_schedule_not_found_fault

See SnapshotScheduleNotFoundFault

snapshot_schedule_quota_exceeded_fault

See SnapshotScheduleQuotaExceededFault

snapshot_schedule_update_in_progress_fault

See SnapshotScheduleUpdateInProgressFault

sns_invalid_topic_fault

See SnsInvalidTopicFault

sns_no_authorization_fault

See SnsNoAuthorizationFault

sns_topic_arn_not_found_fault

See SnsTopicArnNotFoundFault

source_not_found_fault

See SourceNotFoundFault

subnet_already_in_use

See SubnetAlreadyInUse

subscription_already_exist_fault

See SubscriptionAlreadyExistFault

subscription_category_not_found_fault

See SubscriptionCategoryNotFoundFault

subscription_event_id_not_found_fault

See SubscriptionEventIdNotFoundFault

subscription_not_found_fault

See SubscriptionNotFoundFault

subscription_severity_not_found_fault

See SubscriptionSeverityNotFoundFault

table_limit_exceeded_fault

See TableLimitExceededFault

table_restore_not_found_fault

See TableRestoreNotFoundFault

tag_limit_exceeded_fault

See TagLimitExceededFault

unauthorized_operation

See UnauthorizedOperation

unauthorized_partner_integration_fault

See UnauthorizedPartnerIntegrationFault

unknown_snapshot_copy_region_fault

See UnknownSnapshotCopyRegionFault

unsupported_operation_fault

See UnsupportedOperationFault

unsupported_option_fault

See UnsupportedOptionFault

usage_limit_already_exists_fault

See UsageLimitAlreadyExistsFault

usage_limit_not_found_fault

See UsageLimitNotFoundFault

Structs

AcceptReservedNodeExchangeError

AccessToClusterDeniedFault

You are not authorized to access the cluster.

AccessToSnapshotDeniedFault

The owner of the specified snapshot has not authorized your account to access the

AddPartnerError

AssociateDataShareConsumerError

AuthenticationProfileAlreadyExistsFault

The authentication profile already exists.

AuthenticationProfileNotFoundFault

The authentication profile can't be found.

AuthenticationProfileQuotaExceededFault

The size or number of authentication profiles has exceeded the quota.

AuthorizationAlreadyExistsFault

The specified CIDR block or EC2 security group is already authorized for the

AuthorizationNotFoundFault

The specified CIDR IP range or EC2 security group is not authorized for the

AuthorizationQuotaExceededFault

The authorization quota for the cluster security group has been reached.

AuthorizeClusterSecurityGroupIngressError

AuthorizeDataShareError

AuthorizeEndpointAccessError

AuthorizeSnapshotAccessError

BatchDeleteClusterSnapshotsError

BatchDeleteRequestSizeExceededFault

The maximum number for a batch delete of snapshots has been reached. The limit is

BatchModifyClusterSnapshotsError

BatchModifyClusterSnapshotsLimitExceededFault

The maximum number for snapshot identifiers has been reached. The limit is 100.

BucketNotFoundFault

Could not find the specified S3 bucket.

CancelResizeError

ClusterAlreadyExistsFault

The account already has a cluster with the given identifier.

ClusterNotFoundFault

The ClusterIdentifier parameter does not refer to an existing cluster.

ClusterOnLatestRevisionFault

Cluster is already on the latest database revision.

ClusterParameterGroupAlreadyExistsFault

A cluster parameter group with the same name already exists.

ClusterParameterGroupNotFoundFault

The parameter group name does not refer to an existing parameter group.

ClusterParameterGroupQuotaExceededFault

The request would result in the user exceeding the allowed number of cluster

ClusterQuotaExceededFault

The request would exceed the allowed number of cluster instances for this account.

ClusterSecurityGroupAlreadyExistsFault

A cluster security group with the same name already exists.

ClusterSecurityGroupNotFoundFault

The cluster security group name does not refer to an existing cluster security

ClusterSecurityGroupQuotaExceededFault

The request would result in the user exceeding the allowed number of cluster

ClusterSnapshotAlreadyExistsFault

The value specified as a snapshot identifier is already used by an existing

ClusterSnapshotNotFoundFault

The snapshot identifier does not refer to an existing cluster snapshot.

ClusterSnapshotQuotaExceededFault

The request would result in the user exceeding the allowed number of cluster

ClusterSubnetGroupAlreadyExistsFault

A ClusterSubnetGroupName is already used by an existing

ClusterSubnetGroupNotFoundFault

The cluster subnet group name does not refer to an existing cluster subnet

ClusterSubnetGroupQuotaExceededFault

The request would result in user exceeding the allowed number of cluster subnet

ClusterSubnetQuotaExceededFault

The request would result in user exceeding the allowed number of subnets in a

CopyClusterSnapshotError

CopyToRegionDisabledFault

Cross-region snapshot copy was temporarily disabled. Try your request

CreateAuthenticationProfileError

CreateClusterError

CreateClusterParameterGroupError

CreateClusterSecurityGroupError

CreateClusterSnapshotError

CreateClusterSubnetGroupError

CreateEndpointAccessError

CreateEventSubscriptionError

CreateHsmClientCertificateError

CreateHsmConfigurationError

CreateScheduledActionError

CreateSnapshotCopyGrantError

CreateSnapshotScheduleError

CreateTagsError

CreateUsageLimitError

DeauthorizeDataShareError

DeleteAuthenticationProfileError

DeleteClusterError

DeleteClusterParameterGroupError

DeleteClusterSecurityGroupError

DeleteClusterSnapshotError

DeleteClusterSubnetGroupError

DeleteEndpointAccessError

DeleteEventSubscriptionError

DeleteHsmClientCertificateError

DeleteHsmConfigurationError

DeletePartnerError

DeleteScheduledActionError

DeleteSnapshotCopyGrantError

DeleteSnapshotScheduleError

DeleteTagsError

DeleteUsageLimitError

DependentServiceRequestThrottlingFault

The request cannot be completed because a dependent service is throttling requests

DependentServiceUnavailableFault

Your request cannot be completed because a dependent internal service is

DescribeAccountAttributesError

DescribeAuthenticationProfilesError

DescribeClusterDbRevisionsError

DescribeClusterParameterGroupsError

DescribeClusterParametersError

DescribeClusterSecurityGroupsError

DescribeClusterSnapshotsError

DescribeClusterSubnetGroupsError

DescribeClusterTracksError

DescribeClusterVersionsError

DescribeClustersError

DescribeDataSharesError

DescribeDataSharesForConsumerError

DescribeDataSharesForProducerError

DescribeDefaultClusterParametersError

DescribeEndpointAccessError

DescribeEndpointAuthorizationError

DescribeEventCategoriesError

DescribeEventSubscriptionsError

DescribeEventsError

DescribeHsmClientCertificatesError

DescribeHsmConfigurationsError

DescribeLoggingStatusError

DescribeNodeConfigurationOptionsError

DescribeOrderableClusterOptionsError

DescribePartnersError

DescribeReservedNodeOfferingsError

DescribeReservedNodesError

DescribeResizeError

DescribeScheduledActionsError

DescribeSnapshotCopyGrantsError

DescribeSnapshotSchedulesError

DescribeStorageError

DescribeTableRestoreStatusError

DescribeTagsError

DescribeUsageLimitsError

DisableLoggingError

DisableSnapshotCopyError

DisassociateDataShareConsumerError

EnableLoggingError

EnableSnapshotCopyError

EndpointAlreadyExistsFault

The account already has a Redshift-managed VPC endpoint with the given identifier.

EndpointAuthorizationAlreadyExistsFault

The authorization already exists for this endpoint.

EndpointAuthorizationNotFoundFault

The authorization for this endpoint can't be found.

EndpointAuthorizationsPerClusterLimitExceededFault

The number of endpoint authorizations per cluster has exceeded its limit.

EndpointNotFoundFault

The endpoint name doesn't refer to an existing endpoint.

EndpointsPerAuthorizationLimitExceededFault

The number of Redshift-managed VPC endpoints per authorization has exceeded its limit.

EndpointsPerClusterLimitExceededFault

The number of Redshift-managed VPC endpoints per cluster has exceeded its limit.

EventSubscriptionQuotaExceededFault

The request would exceed the allowed number of event subscriptions for this

GetClusterCredentialsError

GetReservedNodeExchangeOfferingsError

HsmClientCertificateAlreadyExistsFault

There is already an existing Amazon Redshift HSM client certificate with the specified

HsmClientCertificateNotFoundFault

There is no Amazon Redshift HSM client certificate with the specified

HsmClientCertificateQuotaExceededFault

The quota for HSM client certificates has been reached.

HsmConfigurationAlreadyExistsFault

There is already an existing Amazon Redshift HSM configuration with the specified

HsmConfigurationNotFoundFault

There is no Amazon Redshift HSM configuration with the specified identifier.

HsmConfigurationQuotaExceededFault

The quota for HSM configurations has been reached.

InProgressTableRestoreQuotaExceededFault

You have exceeded the allowed number of table restore requests. Wait for your

IncompatibleOrderableOptions

The specified options are incompatible.

InsufficientClusterCapacityFault

The number of nodes specified exceeds the allotted capacity of the

InsufficientS3BucketPolicyFault

The cluster does not have read bucket or put object permissions on the S3 bucket

InvalidAuthenticationProfileRequestFault

The authentication profile request is not valid. The profile name can't be null or empty.

InvalidAuthorizationStateFault

The status of the authorization is not valid.

InvalidClusterParameterGroupStateFault

The cluster parameter group action can not be completed because another task is in

InvalidClusterSecurityGroupStateFault

The state of the cluster security group is not available.

InvalidClusterSnapshotScheduleStateFault

The cluster snapshot schedule state is not valid.

InvalidClusterSnapshotStateFault

The specified cluster snapshot is not in the available state, or other

InvalidClusterStateFault

The specified cluster is not in the available state.

InvalidClusterSubnetGroupStateFault

The cluster subnet group cannot be deleted because it is in use.

InvalidClusterSubnetStateFault

The state of the subnet is invalid.

InvalidClusterTrackFault

The provided cluster track name is not valid.

InvalidDataShareFault

There is an error with the datashare.

InvalidElasticIpFault

The Elastic IP (EIP) is invalid or cannot be found.

InvalidEndpointStateFault

The status of the endpoint is not valid.

InvalidHsmClientCertificateStateFault

The specified HSM client certificate is not in the available state, or

InvalidHsmConfigurationStateFault

The specified HSM configuration is not in the available state, or it

InvalidNamespaceFault

The namespace isn't valid because the namespace doesn't exist. Provide a valid namespace.

InvalidReservedNodeStateFault

Indicates that the Reserved Node being exchanged is not in an active state.

InvalidRestoreFault

The restore is invalid.

InvalidRetentionPeriodFault

The retention period specified is either in the past or is not a valid value.

InvalidS3BucketNameFault

The S3 bucket name is invalid. For more information about naming rules, go to

InvalidS3KeyPrefixFault

The string specified for the logging S3 key prefix does not comply with the

InvalidScheduleFault

The schedule you submitted isn't valid.

InvalidScheduledActionFault

The scheduled action is not valid.

InvalidSnapshotCopyGrantStateFault

The snapshot copy grant can't be deleted because it is used by one or more

InvalidSubnet

The requested subnet is not valid, or not all of the subnets are in the same

InvalidSubscriptionStateFault

The subscription request is invalid because it is a duplicate request. This

InvalidTableRestoreArgumentFault

The value specified for the sourceDatabaseName,

InvalidTagFault

The tag is invalid.

InvalidUsageLimitFault

The usage limit is not valid.

InvalidVpcNetworkStateFault

The cluster subnet group does not cover all Availability Zones.

LimitExceededFault

The encryption key has exceeded its grant limit in Amazon Web Services KMS.

ModifyAquaConfigurationError

ModifyAuthenticationProfileError

ModifyClusterDbRevisionError

ModifyClusterError

ModifyClusterIamRolesError

ModifyClusterMaintenanceError

ModifyClusterParameterGroupError

ModifyClusterSnapshotError

ModifyClusterSnapshotScheduleError

ModifyClusterSubnetGroupError

ModifyEndpointAccessError

ModifyEventSubscriptionError

ModifyScheduledActionError

ModifySnapshotCopyRetentionPeriodError

ModifySnapshotScheduleError

ModifyUsageLimitError

NumberOfNodesPerClusterLimitExceededFault

The operation would exceed the number of nodes allowed for a cluster.

NumberOfNodesQuotaExceededFault

The operation would exceed the number of nodes allotted to the account.

PartnerNotFoundFault

The name of the partner was not found.

PauseClusterError

PurchaseReservedNodeOfferingError

RebootClusterError

RejectDataShareError

ReservedNodeAlreadyExistsFault

User already has a reservation with the given identifier.

ReservedNodeAlreadyMigratedFault

Indicates that the reserved node has already been exchanged.

ReservedNodeNotFoundFault

The specified reserved compute node not found.

ReservedNodeOfferingNotFoundFault

Specified offering does not exist.

ReservedNodeQuotaExceededFault

Request would exceed the user's compute node quota.

ResetClusterParameterGroupError

ResizeClusterError

ResizeNotFoundFault

A resize operation for the specified cluster is not found.

ResourceNotFoundFault

The resource could not be found.

RestoreFromClusterSnapshotError

RestoreTableFromClusterSnapshotError

ResumeClusterError

RevokeClusterSecurityGroupIngressError

RevokeEndpointAccessError

RevokeSnapshotAccessError

RotateEncryptionKeyError

ScheduleDefinitionTypeUnsupportedFault

The definition you submitted is not supported.

ScheduledActionAlreadyExistsFault

The scheduled action already exists.

ScheduledActionNotFoundFault

The scheduled action cannot be found.

ScheduledActionQuotaExceededFault

The quota for scheduled actions exceeded.

ScheduledActionTypeUnsupportedFault

The action type specified for a scheduled action is not supported.

SnapshotCopyAlreadyDisabledFault

The cluster already has cross-region snapshot copy disabled.

SnapshotCopyAlreadyEnabledFault

The cluster already has cross-region snapshot copy enabled.

SnapshotCopyDisabledFault

Cross-region snapshot copy was temporarily disabled. Try your request

SnapshotCopyGrantAlreadyExistsFault

The snapshot copy grant can't be created because a grant with the same name already

SnapshotCopyGrantNotFoundFault

The specified snapshot copy grant can't be found. Make sure that the name is typed

SnapshotCopyGrantQuotaExceededFault

The Amazon Web Services account has exceeded the maximum number of snapshot copy grants in this

SnapshotScheduleAlreadyExistsFault

The specified snapshot schedule already exists.

SnapshotScheduleNotFoundFault

We could not find the specified snapshot schedule.

SnapshotScheduleQuotaExceededFault

You have exceeded the quota of snapshot schedules.

SnapshotScheduleUpdateInProgressFault

The specified snapshot schedule is already being updated.

SnsInvalidTopicFault

Amazon SNS has responded that there is a problem with the specified Amazon SNS

SnsNoAuthorizationFault

You do not have permission to publish to the specified Amazon SNS topic.

SnsTopicArnNotFoundFault

An Amazon SNS topic with the specified Amazon Resource Name (ARN) does not

SourceNotFoundFault

The specified Amazon Redshift event source could not be found.

SubnetAlreadyInUse

A specified subnet is already in use by another cluster.

SubscriptionAlreadyExistFault

There is already an existing event notification subscription with the specified

SubscriptionCategoryNotFoundFault

The value specified for the event category was not one of the allowed values, or it

SubscriptionEventIdNotFoundFault

An Amazon Redshift event with the specified event ID does not exist.

SubscriptionNotFoundFault

An Amazon Redshift event notification subscription with the specified name does not

SubscriptionSeverityNotFoundFault

The value specified for the event severity was not one of the allowed values, or it

TableLimitExceededFault

The number of tables in the cluster exceeds the limit for the requested new cluster

TableRestoreNotFoundFault

The specified TableRestoreRequestId value was not found.

TagLimitExceededFault

You have exceeded the number of tags allowed.

UnauthorizedOperation

Your account is not authorized to perform the requested operation.

UnauthorizedPartnerIntegrationFault

The partner integration is not authorized.

UnknownSnapshotCopyRegionFault

The specified region is incorrect or does not exist.

UnsupportedOperationFault

The requested operation isn't supported.

UnsupportedOptionFault

A request option was specified that is not supported.

UpdatePartnerStatusError

UsageLimitAlreadyExistsFault

The usage limit already exists.

UsageLimitNotFoundFault

The usage limit identifier can't be found.

Enums

AcceptReservedNodeExchangeErrorKind

AddPartnerErrorKind

AssociateDataShareConsumerErrorKind

AuthorizeClusterSecurityGroupIngressErrorKind

AuthorizeDataShareErrorKind

AuthorizeEndpointAccessErrorKind

AuthorizeSnapshotAccessErrorKind

BatchDeleteClusterSnapshotsErrorKind

BatchModifyClusterSnapshotsErrorKind

CancelResizeErrorKind

CopyClusterSnapshotErrorKind

CreateAuthenticationProfileErrorKind

CreateClusterErrorKind

CreateClusterParameterGroupErrorKind

CreateClusterSecurityGroupErrorKind

CreateClusterSnapshotErrorKind

CreateClusterSubnetGroupErrorKind

CreateEndpointAccessErrorKind

CreateEventSubscriptionErrorKind

CreateHsmClientCertificateErrorKind

CreateHsmConfigurationErrorKind

CreateScheduledActionErrorKind

CreateSnapshotCopyGrantErrorKind

CreateSnapshotScheduleErrorKind

CreateTagsErrorKind

CreateUsageLimitErrorKind

DeauthorizeDataShareErrorKind

DeleteAuthenticationProfileErrorKind

DeleteClusterErrorKind

DeleteClusterParameterGroupErrorKind

DeleteClusterSecurityGroupErrorKind

DeleteClusterSnapshotErrorKind

DeleteClusterSubnetGroupErrorKind

DeleteEndpointAccessErrorKind

DeleteEventSubscriptionErrorKind

DeleteHsmClientCertificateErrorKind

DeleteHsmConfigurationErrorKind

DeletePartnerErrorKind

DeleteScheduledActionErrorKind

DeleteSnapshotCopyGrantErrorKind

DeleteSnapshotScheduleErrorKind

DeleteTagsErrorKind

DeleteUsageLimitErrorKind

DescribeAccountAttributesErrorKind

DescribeAuthenticationProfilesErrorKind

DescribeClusterDbRevisionsErrorKind

DescribeClusterParameterGroupsErrorKind

DescribeClusterParametersErrorKind

DescribeClusterSecurityGroupsErrorKind

DescribeClusterSnapshotsErrorKind

DescribeClusterSubnetGroupsErrorKind

DescribeClusterTracksErrorKind

DescribeClusterVersionsErrorKind

DescribeClustersErrorKind

DescribeDataSharesErrorKind

DescribeDataSharesForConsumerErrorKind

DescribeDataSharesForProducerErrorKind

DescribeDefaultClusterParametersErrorKind

DescribeEndpointAccessErrorKind

DescribeEndpointAuthorizationErrorKind

DescribeEventCategoriesErrorKind

DescribeEventSubscriptionsErrorKind

DescribeEventsErrorKind

DescribeHsmClientCertificatesErrorKind

DescribeHsmConfigurationsErrorKind

DescribeLoggingStatusErrorKind

DescribeNodeConfigurationOptionsErrorKind

DescribeOrderableClusterOptionsErrorKind

DescribePartnersErrorKind

DescribeReservedNodeOfferingsErrorKind

DescribeReservedNodesErrorKind

DescribeResizeErrorKind

DescribeScheduledActionsErrorKind

DescribeSnapshotCopyGrantsErrorKind

DescribeSnapshotSchedulesErrorKind

DescribeStorageErrorKind

DescribeTableRestoreStatusErrorKind

DescribeTagsErrorKind

DescribeUsageLimitsErrorKind

DisableLoggingErrorKind

DisableSnapshotCopyErrorKind

DisassociateDataShareConsumerErrorKind

EnableLoggingErrorKind

EnableSnapshotCopyErrorKind

GetClusterCredentialsErrorKind

GetReservedNodeExchangeOfferingsErrorKind

ModifyAquaConfigurationErrorKind

ModifyAuthenticationProfileErrorKind

ModifyClusterDbRevisionErrorKind

ModifyClusterErrorKind

ModifyClusterIamRolesErrorKind

ModifyClusterMaintenanceErrorKind

ModifyClusterParameterGroupErrorKind

ModifyClusterSnapshotErrorKind

ModifyClusterSnapshotScheduleErrorKind

ModifyClusterSubnetGroupErrorKind

ModifyEndpointAccessErrorKind

ModifyEventSubscriptionErrorKind

ModifyScheduledActionErrorKind

ModifySnapshotCopyRetentionPeriodErrorKind

ModifySnapshotScheduleErrorKind

ModifyUsageLimitErrorKind

PauseClusterErrorKind

PurchaseReservedNodeOfferingErrorKind

RebootClusterErrorKind

RejectDataShareErrorKind

ResetClusterParameterGroupErrorKind

ResizeClusterErrorKind

RestoreFromClusterSnapshotErrorKind

RestoreTableFromClusterSnapshotErrorKind

ResumeClusterErrorKind

RevokeClusterSecurityGroupIngressErrorKind

RevokeEndpointAccessErrorKind

RevokeSnapshotAccessErrorKind

RotateEncryptionKeyErrorKind

UpdatePartnerStatusErrorKind