Module aws_sdk_redshift::input

Modules

accept_reserved_node_exchange_input

See AcceptReservedNodeExchangeInput

add_partner_input

See AddPartnerInput

associate_data_share_consumer_input

See AssociateDataShareConsumerInput

authorize_cluster_security_group_ingress_input

See AuthorizeClusterSecurityGroupIngressInput

authorize_data_share_input

See AuthorizeDataShareInput

authorize_endpoint_access_input

See AuthorizeEndpointAccessInput

authorize_snapshot_access_input

See AuthorizeSnapshotAccessInput

batch_delete_cluster_snapshots_input

See BatchDeleteClusterSnapshotsInput

batch_modify_cluster_snapshots_input

See BatchModifyClusterSnapshotsInput

cancel_resize_input

See CancelResizeInput

copy_cluster_snapshot_input

See CopyClusterSnapshotInput

create_authentication_profile_input

See CreateAuthenticationProfileInput

create_cluster_input

See CreateClusterInput

create_cluster_parameter_group_input

See CreateClusterParameterGroupInput

create_cluster_security_group_input

See CreateClusterSecurityGroupInput

create_cluster_snapshot_input

See CreateClusterSnapshotInput

create_cluster_subnet_group_input

See CreateClusterSubnetGroupInput

create_endpoint_access_input

See CreateEndpointAccessInput

create_event_subscription_input

See CreateEventSubscriptionInput

create_hsm_client_certificate_input

See CreateHsmClientCertificateInput

create_hsm_configuration_input

See CreateHsmConfigurationInput

create_scheduled_action_input

See CreateScheduledActionInput

create_snapshot_copy_grant_input

See CreateSnapshotCopyGrantInput

create_snapshot_schedule_input

See CreateSnapshotScheduleInput

create_tags_input

See CreateTagsInput

create_usage_limit_input

See CreateUsageLimitInput

deauthorize_data_share_input

See DeauthorizeDataShareInput

delete_authentication_profile_input

See DeleteAuthenticationProfileInput

delete_cluster_input

See DeleteClusterInput

delete_cluster_parameter_group_input

See DeleteClusterParameterGroupInput

delete_cluster_security_group_input

See DeleteClusterSecurityGroupInput

delete_cluster_snapshot_input

See DeleteClusterSnapshotInput

delete_cluster_subnet_group_input

See DeleteClusterSubnetGroupInput

delete_endpoint_access_input

See DeleteEndpointAccessInput

delete_event_subscription_input

See DeleteEventSubscriptionInput

delete_hsm_client_certificate_input

See DeleteHsmClientCertificateInput

delete_hsm_configuration_input

See DeleteHsmConfigurationInput

delete_partner_input

See DeletePartnerInput

delete_scheduled_action_input

See DeleteScheduledActionInput

delete_snapshot_copy_grant_input

See DeleteSnapshotCopyGrantInput

delete_snapshot_schedule_input

See DeleteSnapshotScheduleInput

delete_tags_input

See DeleteTagsInput

delete_usage_limit_input

See DeleteUsageLimitInput

describe_account_attributes_input

See DescribeAccountAttributesInput

describe_authentication_profiles_input

See DescribeAuthenticationProfilesInput

describe_cluster_db_revisions_input

See DescribeClusterDbRevisionsInput

describe_cluster_parameter_groups_input

See DescribeClusterParameterGroupsInput

describe_cluster_parameters_input

See DescribeClusterParametersInput

describe_cluster_security_groups_input

See DescribeClusterSecurityGroupsInput

describe_cluster_snapshots_input

See DescribeClusterSnapshotsInput

describe_cluster_subnet_groups_input

See DescribeClusterSubnetGroupsInput

describe_cluster_tracks_input

See DescribeClusterTracksInput

describe_cluster_versions_input

See DescribeClusterVersionsInput

describe_clusters_input

See DescribeClustersInput

describe_data_shares_for_consumer_input

See DescribeDataSharesForConsumerInput

describe_data_shares_for_producer_input

See DescribeDataSharesForProducerInput

describe_data_shares_input

See DescribeDataSharesInput

describe_default_cluster_parameters_input

See DescribeDefaultClusterParametersInput

describe_endpoint_access_input

See DescribeEndpointAccessInput

describe_endpoint_authorization_input

See DescribeEndpointAuthorizationInput

describe_event_categories_input

See DescribeEventCategoriesInput

describe_event_subscriptions_input

See DescribeEventSubscriptionsInput

describe_events_input

See DescribeEventsInput

describe_hsm_client_certificates_input

See DescribeHsmClientCertificatesInput

describe_hsm_configurations_input

See DescribeHsmConfigurationsInput

describe_logging_status_input

See DescribeLoggingStatusInput

describe_node_configuration_options_input

See DescribeNodeConfigurationOptionsInput

describe_orderable_cluster_options_input

See DescribeOrderableClusterOptionsInput

describe_partners_input

See DescribePartnersInput

describe_reserved_node_offerings_input

See DescribeReservedNodeOfferingsInput

describe_reserved_nodes_input

See DescribeReservedNodesInput

describe_resize_input

See DescribeResizeInput

describe_scheduled_actions_input

See DescribeScheduledActionsInput

describe_snapshot_copy_grants_input

See DescribeSnapshotCopyGrantsInput

describe_snapshot_schedules_input

See DescribeSnapshotSchedulesInput

describe_storage_input

See DescribeStorageInput

describe_table_restore_status_input

See DescribeTableRestoreStatusInput

describe_tags_input

See DescribeTagsInput

describe_usage_limits_input

See DescribeUsageLimitsInput

disable_logging_input

See DisableLoggingInput

disable_snapshot_copy_input

See DisableSnapshotCopyInput

disassociate_data_share_consumer_input

See DisassociateDataShareConsumerInput

enable_logging_input

See EnableLoggingInput

enable_snapshot_copy_input

See EnableSnapshotCopyInput

get_cluster_credentials_input

See GetClusterCredentialsInput

get_reserved_node_exchange_offerings_input

See GetReservedNodeExchangeOfferingsInput

modify_aqua_configuration_input

See ModifyAquaConfigurationInput

modify_authentication_profile_input

See ModifyAuthenticationProfileInput

modify_cluster_db_revision_input

See ModifyClusterDbRevisionInput

modify_cluster_iam_roles_input

See ModifyClusterIamRolesInput

modify_cluster_input

See ModifyClusterInput

modify_cluster_maintenance_input

See ModifyClusterMaintenanceInput

modify_cluster_parameter_group_input

See ModifyClusterParameterGroupInput

modify_cluster_snapshot_input

See ModifyClusterSnapshotInput

modify_cluster_snapshot_schedule_input

See ModifyClusterSnapshotScheduleInput

modify_cluster_subnet_group_input

See ModifyClusterSubnetGroupInput

modify_endpoint_access_input

See ModifyEndpointAccessInput

modify_event_subscription_input

See ModifyEventSubscriptionInput

modify_scheduled_action_input

See ModifyScheduledActionInput

modify_snapshot_copy_retention_period_input

See ModifySnapshotCopyRetentionPeriodInput

modify_snapshot_schedule_input

See ModifySnapshotScheduleInput

modify_usage_limit_input

See ModifyUsageLimitInput

pause_cluster_input

See PauseClusterInput

purchase_reserved_node_offering_input

See PurchaseReservedNodeOfferingInput

reboot_cluster_input

See RebootClusterInput

reject_data_share_input

See RejectDataShareInput

reset_cluster_parameter_group_input

See ResetClusterParameterGroupInput

resize_cluster_input

See ResizeClusterInput

restore_from_cluster_snapshot_input

See RestoreFromClusterSnapshotInput

restore_table_from_cluster_snapshot_input

See RestoreTableFromClusterSnapshotInput

resume_cluster_input

See ResumeClusterInput

revoke_cluster_security_group_ingress_input

See RevokeClusterSecurityGroupIngressInput

revoke_endpoint_access_input

See RevokeEndpointAccessInput

revoke_snapshot_access_input

See RevokeSnapshotAccessInput

rotate_encryption_key_input

See RotateEncryptionKeyInput

update_partner_status_input

See UpdatePartnerStatusInput

Structs

AcceptReservedNodeExchangeInput

AddPartnerInput

AssociateDataShareConsumerInput

AuthorizeClusterSecurityGroupIngressInput

AuthorizeDataShareInput

AuthorizeEndpointAccessInput

AuthorizeSnapshotAccessInput

BatchDeleteClusterSnapshotsInput

BatchModifyClusterSnapshotsInput

CancelResizeInput

CopyClusterSnapshotInput

CreateAuthenticationProfileInput

CreateClusterInput

CreateClusterParameterGroupInput

CreateClusterSecurityGroupInput

CreateClusterSnapshotInput

CreateClusterSubnetGroupInput

CreateEndpointAccessInput

CreateEventSubscriptionInput

CreateHsmClientCertificateInput

CreateHsmConfigurationInput

CreateScheduledActionInput

CreateSnapshotCopyGrantInput

The result of the CreateSnapshotCopyGrant action.

CreateSnapshotScheduleInput

CreateTagsInput

Contains the output from the CreateTags action.

CreateUsageLimitInput

DeauthorizeDataShareInput

DeleteAuthenticationProfileInput

DeleteClusterInput

DeleteClusterParameterGroupInput

DeleteClusterSecurityGroupInput

DeleteClusterSnapshotInput

DeleteClusterSubnetGroupInput

DeleteEndpointAccessInput

DeleteEventSubscriptionInput

DeleteHsmClientCertificateInput

DeleteHsmConfigurationInput

DeletePartnerInput

DeleteScheduledActionInput

DeleteSnapshotCopyGrantInput

The result of the DeleteSnapshotCopyGrant action.

DeleteSnapshotScheduleInput

DeleteTagsInput

Contains the output from the DeleteTags action.

DeleteUsageLimitInput

DescribeAccountAttributesInput

DescribeAuthenticationProfilesInput

DescribeClusterDbRevisionsInput

DescribeClusterParameterGroupsInput

DescribeClusterParametersInput

DescribeClusterSecurityGroupsInput

DescribeClusterSnapshotsInput

DescribeClusterSubnetGroupsInput

DescribeClusterTracksInput

DescribeClusterVersionsInput

DescribeClustersInput

DescribeDataSharesForConsumerInput

DescribeDataSharesForProducerInput

DescribeDataSharesInput

DescribeDefaultClusterParametersInput

DescribeEndpointAccessInput

DescribeEndpointAuthorizationInput

DescribeEventCategoriesInput

DescribeEventSubscriptionsInput

DescribeEventsInput

DescribeHsmClientCertificatesInput

DescribeHsmConfigurationsInput

DescribeLoggingStatusInput

DescribeNodeConfigurationOptionsInput

DescribeOrderableClusterOptionsInput

DescribePartnersInput

DescribeReservedNodeOfferingsInput

DescribeReservedNodesInput

DescribeResizeInput

DescribeScheduledActionsInput

DescribeSnapshotCopyGrantsInput

The result of the DescribeSnapshotCopyGrants action.

DescribeSnapshotSchedulesInput

DescribeStorageInput

DescribeTableRestoreStatusInput

DescribeTagsInput

DescribeUsageLimitsInput

DisableLoggingInput

DisableSnapshotCopyInput

DisassociateDataShareConsumerInput

EnableLoggingInput

EnableSnapshotCopyInput

GetClusterCredentialsInput

The request parameters to get cluster credentials.

GetReservedNodeExchangeOfferingsInput

ModifyAquaConfigurationInput

ModifyAuthenticationProfileInput

ModifyClusterDbRevisionInput

ModifyClusterIamRolesInput

ModifyClusterInput

ModifyClusterMaintenanceInput

ModifyClusterParameterGroupInput

Describes a modify cluster parameter group operation.

ModifyClusterSnapshotInput

ModifyClusterSnapshotScheduleInput

ModifyClusterSubnetGroupInput

ModifyEndpointAccessInput

ModifyEventSubscriptionInput

ModifyScheduledActionInput

ModifySnapshotCopyRetentionPeriodInput

ModifySnapshotScheduleInput

ModifyUsageLimitInput

PauseClusterInput

Describes a pause cluster operation. For example, a scheduled action to run the PauseCluster API operation.

PurchaseReservedNodeOfferingInput

RebootClusterInput

RejectDataShareInput

ResetClusterParameterGroupInput

ResizeClusterInput

Describes a resize cluster operation. For example, a scheduled action to run the ResizeCluster API operation.

RestoreFromClusterSnapshotInput

RestoreTableFromClusterSnapshotInput

ResumeClusterInput

Describes a resume cluster operation. For example, a scheduled action to run the ResumeCluster API operation.

RevokeClusterSecurityGroupIngressInput

RevokeEndpointAccessInput

RevokeSnapshotAccessInput

RotateEncryptionKeyInput

UpdatePartnerStatusInput