Module aws_sdk_redshift::model

Modules

account_attribute

See AccountAttribute

account_with_restore_access

See AccountWithRestoreAccess

aqua_configuration

See AquaConfiguration

attribute_value_target

See AttributeValueTarget

authentication_profile

See AuthenticationProfile

availability_zone

See AvailabilityZone

cluster

See Cluster

cluster_associated_to_schedule

See ClusterAssociatedToSchedule

cluster_db_revision

See ClusterDbRevision

cluster_iam_role

See ClusterIamRole

cluster_node

See ClusterNode

cluster_parameter_group

See ClusterParameterGroup

cluster_parameter_group_status

See ClusterParameterGroupStatus

cluster_parameter_status

See ClusterParameterStatus

cluster_security_group

See ClusterSecurityGroup

cluster_security_group_membership

See ClusterSecurityGroupMembership

cluster_snapshot_copy_status

See ClusterSnapshotCopyStatus

cluster_subnet_group

See ClusterSubnetGroup

cluster_version

See ClusterVersion

data_share

See DataShare

data_share_association

See DataShareAssociation

data_transfer_progress

See DataTransferProgress

default_cluster_parameters

See DefaultClusterParameters

deferred_maintenance_window

See DeferredMaintenanceWindow

delete_cluster_snapshot_message

See DeleteClusterSnapshotMessage

ec2_security_group

See Ec2SecurityGroup

elastic_ip_status

See ElasticIpStatus

endpoint

See Endpoint

endpoint_access

See EndpointAccess

endpoint_authorization

See EndpointAuthorization

event

See Event

event_categories_map

See EventCategoriesMap

event_info_map

See EventInfoMap

event_subscription

See EventSubscription

hsm_client_certificate

See HsmClientCertificate

hsm_configuration

See HsmConfiguration

hsm_status

See HsmStatus

ip_range

See IpRange

maintenance_track

See MaintenanceTrack

network_interface

See NetworkInterface

node_configuration_option

See NodeConfigurationOption

node_configuration_options_filter

See NodeConfigurationOptionsFilter

orderable_cluster_option

See OrderableClusterOption

parameter

See Parameter

partner_integration_info

See PartnerIntegrationInfo

pause_cluster_message

See PauseClusterMessage

pending_modified_values

See PendingModifiedValues

recurring_charge

See RecurringCharge

reserved_node

See ReservedNode

reserved_node_offering

See ReservedNodeOffering

resize_cluster_message

See ResizeClusterMessage

resize_info

See ResizeInfo

restore_status

See RestoreStatus

resume_cluster_message

See ResumeClusterMessage

revision_target

See RevisionTarget

scheduled_action

See ScheduledAction

scheduled_action_filter

See ScheduledActionFilter

scheduled_action_type

See ScheduledActionType

snapshot

See Snapshot

snapshot_copy_grant

See SnapshotCopyGrant

snapshot_error_message

See SnapshotErrorMessage

snapshot_schedule

See SnapshotSchedule

snapshot_sorting_entity

See SnapshotSortingEntity

subnet

See Subnet

supported_operation

See SupportedOperation

supported_platform

See SupportedPlatform

table_restore_status

See TableRestoreStatus

tag

See Tag

tagged_resource

See TaggedResource

update_target

See UpdateTarget

usage_limit

See UsageLimit

vpc_endpoint

See VpcEndpoint

vpc_security_group_membership

See VpcSecurityGroupMembership

Structs

AccountAttribute

A name value pair that describes an aspect of an account.

AccountWithRestoreAccess

Describes an Amazon Web Services account authorized to restore a snapshot.

AquaConfiguration

The AQUA (Advanced Query Accelerator) configuration of the cluster.

AttributeValueTarget

Describes an attribute value.

AuthenticationProfile

Describes an authentication profile.

AvailabilityZone

Describes an availability zone.

Cluster

Describes a cluster.

ClusterAssociatedToSchedule

ClusterDbRevision

Describes a ClusterDbRevision.

ClusterIamRole

An Identity and Access Management (IAM) role that can be used by the associated

ClusterNode

The identifier of a node in a cluster.

ClusterParameterGroup

Describes a parameter group.

ClusterParameterGroupStatus

Describes the status of a parameter group.

ClusterParameterStatus

Describes the status of a parameter group.

ClusterSecurityGroup

Describes a security group.

ClusterSecurityGroupMembership

Describes a cluster security group.

ClusterSnapshotCopyStatus

Returns the destination region and retention period that are configured for

ClusterSubnetGroup

Describes a subnet group.

ClusterVersion

Describes a cluster version, including the parameter group family and description

DataShare

DataShareAssociation

The association of a datashare from a producer account with a data consumer.

DataTransferProgress

Describes the status of a cluster while it is in the process of resizing with an

DefaultClusterParameters

Describes the default cluster parameters for a parameter group family.

DeferredMaintenanceWindow

Describes a deferred maintenance window

DeleteClusterSnapshotMessage

Ec2SecurityGroup

Describes an Amazon EC2 security group.

ElasticIpStatus

Describes the status of the elastic IP (EIP) address.

Endpoint

Describes a connection endpoint.

EndpointAccess

Describes a Redshift-managed VPC endpoint.

EndpointAuthorization

Describes an endpoint authorization for authorizing Redshift-managed VPC endpoint access to a cluster across Amazon Web Services accounts.

Event

Describes an event.

EventCategoriesMap

Describes event categories.

EventInfoMap

Describes event information.

EventSubscription

Describes event subscriptions.

HsmClientCertificate

Returns information about an HSM client certificate. The certificate is stored in a

HsmConfiguration

Returns information about an HSM configuration, which is an object that describes

HsmStatus

Describes the status of changes to HSM settings.

IpRange

Describes an IP range used in a security group.

MaintenanceTrack

Defines a maintenance track that determines which Amazon Redshift version to apply

NetworkInterface

Describes a network interface.

NodeConfigurationOption

A list of node configurations.

NodeConfigurationOptionsFilter

A set of elements to filter the returned node configurations.

OrderableClusterOption

Describes an orderable cluster option.

Parameter

Describes a parameter in a cluster parameter group.

PartnerIntegrationInfo

Describes a partner integration.

PauseClusterMessage

Describes a pause cluster operation. For example, a scheduled action to run the PauseCluster API operation.

PendingModifiedValues

Describes cluster attributes that are in a pending state. A change to one or more

RecurringCharge

Describes a recurring charge.

ReservedNode

Describes a reserved node. You can call the DescribeReservedNodeOfferings API to obtain the available reserved node

ReservedNodeOffering

Describes a reserved node offering.

ResizeClusterMessage

Describes a resize cluster operation. For example, a scheduled action to run the ResizeCluster API operation.

ResizeInfo

Describes a resize operation.

RestoreStatus

Describes the status of a cluster restore action. Returns null if the cluster was

ResumeClusterMessage

Describes a resume cluster operation. For example, a scheduled action to run the ResumeCluster API operation.

RevisionTarget

Describes a RevisionTarget.

ScheduledAction

Describes a scheduled action. You can use a scheduled action to trigger some Amazon Redshift API operations on a schedule.

ScheduledActionFilter

A set of elements to filter the returned scheduled actions.

ScheduledActionType

The action type that specifies an Amazon Redshift API operation that is supported by the Amazon Redshift scheduler.

Snapshot

Describes a snapshot.

SnapshotCopyGrant

The snapshot copy grant that grants Amazon Redshift permission to encrypt copied

SnapshotErrorMessage

Describes the errors returned by a snapshot.

SnapshotSchedule

Describes a snapshot schedule. You can set a regular interval for creating

SnapshotSortingEntity

Describes a sorting entity

Subnet

Describes a subnet.

SupportedOperation

Describes the operations that are allowed on a maintenance track.

SupportedPlatform

A list of supported platforms for orderable clusters.

TableRestoreStatus

Describes the status of a RestoreTableFromClusterSnapshot

Tag

A tag consisting of a name/value pair for a resource.

TaggedResource

A tag and its associated resource.

UpdateTarget

A maintenance track that you can switch the current track to.

UsageLimit

Describes a usage limit object for a cluster.

VpcEndpoint

The connection endpoint for connecting to an Amazon Redshift cluster through the proxy.

VpcSecurityGroupMembership

Describes the members of a VPC security group.

Enums

ActionType

AquaConfigurationStatus

AquaStatus

AuthorizationStatus

DataShareStatus

DataShareStatusForConsumer

DataShareStatusForProducer

Mode

NodeConfigurationOptionsFilterName

OperatorType

ParameterApplyType

PartnerIntegrationStatus

ReservedNodeOfferingType

ScheduleState

ScheduledActionFilterName

ScheduledActionState

ScheduledActionTypeValues

SnapshotAttributeToSortBy

SortByOrder

SourceType

TableRestoreStatusType

UsageLimitBreachAction

UsageLimitFeatureType

UsageLimitLimitType

UsageLimitPeriod