Struct aws_sdk_redshift::model::cluster::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Cluster

Implementations

impl Builder

pub fn cluster_identifier(self, input: impl Into<String>) -> Self

The unique identifier of the cluster.

pub fn set_cluster_identifier(self, input: Option<String>) -> Self

pub fn node_type(self, input: impl Into<String>) -> Self

The node type for the nodes in the cluster.

pub fn set_node_type(self, input: Option<String>) -> Self

pub fn cluster_status(self, input: impl Into<String>) -> Self

The current state of the cluster. Possible values are the following:

available

available, prep-for-resize

available, resize-cleanup

cancelling-resize

creating

deleting

final-snapshot

hardware-failure

incompatible-hsm

incompatible-network

incompatible-parameters

incompatible-restore

modifying

paused

rebooting

renaming

resizing

rotating-keys

storage-full

updating-hsm

pub fn set_cluster_status(self, input: Option<String>) -> Self

pub fn cluster_availability_status(self, input: impl Into<String>) -> Self

The availability status of the cluster for queries. Possible values are the following:

Available - The cluster is available for queries.

Unavailable - The cluster is not available for queries.

Maintenance - The cluster is intermittently available for queries due to maintenance activities.

Modifying - The cluster is intermittently available for queries due to changes that modify the cluster.

Failed - The cluster failed and is not available for queries.

pub fn set_cluster_availability_status(self, input: Option<String>) -> Self

pub fn modify_status(self, input: impl Into<String>) -> Self

The status of a modify operation, if any, initiated for the cluster.

pub fn set_modify_status(self, input: Option<String>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

The admin user name for the cluster. This name is used to connect to the database that is specified in the DBName parameter.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn db_name(self, input: impl Into<String>) -> Self

The name of the initial database that was created when the cluster was created. This same name is returned for the life of the cluster. If an initial database was not specified, a database named devdev was created by default.

pub fn set_db_name(self, input: Option<String>) -> Self

pub fn endpoint(self, input: Endpoint) -> Self

The connection endpoint.

pub fn set_endpoint(self, input: Option<Endpoint>) -> Self

pub fn cluster_create_time(self, input: Instant) -> Self

The date and time that the cluster was created.

pub fn set_cluster_create_time(self, input: Option<Instant>) -> Self

pub fn automated_snapshot_retention_period(self, input: i32) -> Self

The number of days that automatic cluster snapshots are retained.

pub fn set_automated_snapshot_retention_period(self, input: Option<i32>) -> Self

pub fn manual_snapshot_retention_period(self, input: i32) -> Self

The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

pub fn set_manual_snapshot_retention_period(self, input: Option<i32>) -> Self

pub fn cluster_security_groups(
self,
input: impl Into<ClusterSecurityGroupMembership>
) -> Self

pub fn set_cluster_security_groups(
self,
input: Option<Vec<ClusterSecurityGroupMembership>>
) -> Self

pub fn vpc_security_groups(
self,
input: impl Into<VpcSecurityGroupMembership>
) -> Self

pub fn set_vpc_security_groups(
self,
input: Option<Vec<VpcSecurityGroupMembership>>
) -> Self

pub fn cluster_parameter_groups(
self,
input: impl Into<ClusterParameterGroupStatus>
) -> Self

pub fn set_cluster_parameter_groups(
self,
input: Option<Vec<ClusterParameterGroupStatus>>
) -> Self

pub fn cluster_subnet_group_name(self, input: impl Into<String>) -> Self

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

pub fn set_cluster_subnet_group_name(self, input: Option<String>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The identifier of the VPC the cluster is in, if the cluster is in a VPC.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The name of the Availability Zone in which the cluster is located.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The weekly time range, in Universal Coordinated Time (UTC), during which system maintenance can occur.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn pending_modified_values(self, input: PendingModifiedValues) -> Self

A value that, if present, indicates that changes to the cluster are pending. Specific pending changes are identified by subelements.

pub fn set_pending_modified_values(
self,
input: Option<PendingModifiedValues>
) -> Self

pub fn cluster_version(self, input: impl Into<String>) -> Self

The version ID of the Amazon Redshift engine that is running on the cluster.

pub fn set_cluster_version(self, input: Option<String>) -> Self

pub fn allow_version_upgrade(self, input: bool) -> Self

A boolean value that, if true, indicates that major version upgrades will be applied automatically to the cluster during the maintenance window.

pub fn set_allow_version_upgrade(self, input: Option<bool>) -> Self

pub fn number_of_nodes(self, input: i32) -> Self

The number of compute nodes in the cluster.

pub fn set_number_of_nodes(self, input: Option<i32>) -> Self

pub fn publicly_accessible(self, input: bool) -> Self

A boolean value that, if true, indicates that the cluster can be accessed from a public network.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn encrypted(self, input: bool) -> Self

A boolean value that, if true, indicates that data in the cluster is encrypted at rest.

pub fn set_encrypted(self, input: Option<bool>) -> Self

pub fn restore_status(self, input: RestoreStatus) -> Self

A value that describes the status of a cluster restore action. This parameter returns null if the cluster was not created by restoring a snapshot.

pub fn set_restore_status(self, input: Option<RestoreStatus>) -> Self

pub fn data_transfer_progress(self, input: DataTransferProgress) -> Self

pub fn set_data_transfer_progress(
self,
input: Option<DataTransferProgress>
) -> Self

pub fn hsm_status(self, input: HsmStatus) -> Self

A value that reports whether the Amazon Redshift cluster has finished applying any hardware security module (HSM) settings changes specified in a modify cluster command.

Values: active, applying

pub fn set_hsm_status(self, input: Option<HsmStatus>) -> Self

pub fn cluster_snapshot_copy_status(
self,
input: ClusterSnapshotCopyStatus
) -> Self

A value that returns the destination region and retention period that are configured for cross-region snapshot copy.

pub fn set_cluster_snapshot_copy_status(
self,
input: Option<ClusterSnapshotCopyStatus>
) -> Self

pub fn cluster_public_key(self, input: impl Into<String>) -> Self

The public key for the cluster.

pub fn set_cluster_public_key(self, input: Option<String>) -> Self

pub fn cluster_nodes(self, input: impl Into<ClusterNode>) -> Self

pub fn set_cluster_nodes(self, input: Option<Vec<ClusterNode>>) -> Self

pub fn elastic_ip_status(self, input: ElasticIpStatus) -> Self

The status of the elastic IP (EIP) address.

pub fn set_elastic_ip_status(self, input: Option<ElasticIpStatus>) -> Self

pub fn cluster_revision_number(self, input: impl Into<String>) -> Self

The specific revision number of the database in the cluster.

pub fn set_cluster_revision_number(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Key Management Service (KMS) key ID of the encryption key used to encrypt data in the cluster.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn enhanced_vpc_routing(self, input: bool) -> Self

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

pub fn set_enhanced_vpc_routing(self, input: Option<bool>) -> Self

pub fn iam_roles(self, input: impl Into<ClusterIamRole>) -> Self

pub fn set_iam_roles(self, input: Option<Vec<ClusterIamRole>>) -> Self

pub fn pending_actions(self, input: impl Into<String>) -> Self

pub fn set_pending_actions(self, input: Option<Vec<String>>) -> Self

pub fn maintenance_track_name(self, input: impl Into<String>) -> Self

The name of the maintenance track for the cluster.

pub fn set_maintenance_track_name(self, input: Option<String>) -> Self

pub fn elastic_resize_number_of_node_options(
self,
input: impl Into<String>
) -> Self

The number of nodes that you can resize the cluster to with the elastic resize method.

pub fn set_elastic_resize_number_of_node_options(
self,
input: Option<String>
) -> Self

pub fn deferred_maintenance_windows(
self,
input: impl Into<DeferredMaintenanceWindow>
) -> Self

pub fn set_deferred_maintenance_windows(
self,
input: Option<Vec<DeferredMaintenanceWindow>>
) -> Self

pub fn snapshot_schedule_identifier(self, input: impl Into<String>) -> Self

A unique identifier for the cluster snapshot schedule.

pub fn set_snapshot_schedule_identifier(self, input: Option<String>) -> Self

pub fn snapshot_schedule_state(self, input: ScheduleState) -> Self

The current state of the cluster snapshot schedule.

pub fn set_snapshot_schedule_state(self, input: Option<ScheduleState>) -> Self

pub fn expected_next_snapshot_schedule_time(self, input: Instant) -> Self

The date and time when the next snapshot is expected to be taken for clusters with a valid snapshot schedule and backups enabled.

pub fn set_expected_next_snapshot_schedule_time(
self,
input: Option<Instant>
) -> Self

pub fn expected_next_snapshot_schedule_time_status(
self,
input: impl Into<String>
) -> Self

The status of next expected snapshot for clusters having a valid snapshot schedule and backups enabled. Possible values are the following:

OnTrack - The next snapshot is expected to be taken on time.

Pending - The next snapshot is pending to be taken.

pub fn set_expected_next_snapshot_schedule_time_status(
self,
input: Option<String>
) -> Self

pub fn next_maintenance_window_start_time(self, input: Instant) -> Self

The date and time in UTC when system maintenance can begin.

pub fn set_next_maintenance_window_start_time(
self,
input: Option<Instant>
) -> Self

pub fn resize_info(self, input: ResizeInfo) -> Self

Returns the following:

AllowCancelResize: a boolean value indicating if the resize operation can be cancelled.

ResizeType: Returns ClassicResize

pub fn set_resize_info(self, input: Option<ResizeInfo>) -> Self

pub fn availability_zone_relocation_status(
self,
input: impl Into<String>
) -> Self

Describes the status of the Availability Zone relocation operation.

pub fn set_availability_zone_relocation_status(
self,
input: Option<String>
) -> Self

pub fn cluster_namespace_arn(self, input: impl Into<String>) -> Self

The namespace Amazon Resource Name (ARN) of the cluster.

pub fn set_cluster_namespace_arn(self, input: Option<String>) -> Self

pub fn total_storage_capacity_in_mega_bytes(self, input: i64) -> Self

The total storage capacity of the cluster in megabytes.

pub fn set_total_storage_capacity_in_mega_bytes(
self,
input: Option<i64>
) -> Self

pub fn aqua_configuration(self, input: AquaConfiguration) -> Self

The AQUA (Advanced Query Accelerator) configuration of the cluster.

pub fn set_aqua_configuration(self, input: Option<AquaConfiguration>) -> Self

pub fn build(self) -> Cluster

Consumes the builder and constructs a Cluster