Struct aws_sdk_redshift::model::snapshot::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Snapshot

Implementations

impl Builder

pub fn snapshot_identifier(self, input: impl Into<String>) -> Self

The snapshot identifier that is provided in the request.

pub fn set_snapshot_identifier(self, input: Option<String>) -> Self

pub fn cluster_identifier(self, input: impl Into<String>) -> Self

The identifier of the cluster for which the snapshot was taken.

pub fn set_cluster_identifier(self, input: Option<String>) -> Self

pub fn snapshot_create_time(self, input: Instant) -> Self

The time (in UTC format) when Amazon Redshift began the snapshot. A snapshot contains a copy of the cluster data as of this exact time.

pub fn set_snapshot_create_time(self, input: Option<Instant>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The snapshot status. The value of the status depends on the API operation used:

CreateClusterSnapshot and CopyClusterSnapshot returns status as "creating".

DescribeClusterSnapshots returns status as "creating", "available", "final snapshot", or "failed".

DeleteClusterSnapshot returns status as "deleted".

pub fn set_status(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port that the cluster is listening on.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone in which the cluster was created.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn cluster_create_time(self, input: Instant) -> Self

The time (UTC) when the cluster was originally created.

pub fn set_cluster_create_time(self, input: Option<Instant>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

The admin user name for the cluster.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn cluster_version(self, input: impl Into<String>) -> Self

The version ID of the Amazon Redshift engine that is running on the cluster.

pub fn set_cluster_version(self, input: Option<String>) -> Self

pub fn engine_full_version(self, input: impl Into<String>) -> Self

The cluster version of the cluster used to create the snapshot. For example, 1.0.15503.

pub fn set_engine_full_version(self, input: Option<String>) -> Self

pub fn snapshot_type(self, input: impl Into<String>) -> Self

The snapshot type. Snapshots created using CreateClusterSnapshot and CopyClusterSnapshot are of type "manual".

pub fn set_snapshot_type(self, input: Option<String>) -> Self

pub fn node_type(self, input: impl Into<String>) -> Self

The node type of the nodes in the cluster.

pub fn set_node_type(self, input: Option<String>) -> Self

pub fn number_of_nodes(self, input: i32) -> Self

The number of nodes in the cluster.

pub fn set_number_of_nodes(self, input: Option<i32>) -> Self

pub fn db_name(self, input: impl Into<String>) -> Self

The name of the database that was created when the cluster was created.

pub fn set_db_name(self, input: Option<String>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The VPC identifier of the cluster if the snapshot is from a cluster in a VPC. Otherwise, this field is not in the output.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn encrypted(self, input: bool) -> Self

If true, the data in the snapshot is encrypted at rest.

pub fn set_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Key Management Service (KMS) key ID of the encryption key that was used to encrypt data in the cluster from which the snapshot was taken.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn encrypted_with_hsm(self, input: bool) -> Self

A boolean that indicates whether the snapshot data is encrypted using the HSM keys of the source cluster. true indicates that the data is encrypted using HSM keys.

pub fn set_encrypted_with_hsm(self, input: Option<bool>) -> Self

pub fn accounts_with_restore_access(
self,
input: impl Into<AccountWithRestoreAccess>
) -> Self

pub fn set_accounts_with_restore_access(
self,
input: Option<Vec<AccountWithRestoreAccess>>
) -> Self

pub fn owner_account(self, input: impl Into<String>) -> Self

For manual snapshots, the Amazon Web Services account used to create or copy the snapshot. For automatic snapshots, the owner of the cluster. The owner can perform all snapshot actions, such as sharing a manual snapshot.

pub fn set_owner_account(self, input: Option<String>) -> Self

pub fn total_backup_size_in_mega_bytes(self, input: f64) -> Self

The size of the complete set of backup data that would be used to restore the cluster.

pub fn set_total_backup_size_in_mega_bytes(self, input: Option<f64>) -> Self

pub fn actual_incremental_backup_size_in_mega_bytes(self, input: f64) -> Self

The size of the incremental backup.

pub fn set_actual_incremental_backup_size_in_mega_bytes(
self,
input: Option<f64>
) -> Self

pub fn backup_progress_in_mega_bytes(self, input: f64) -> Self

The number of megabytes that have been transferred to the snapshot backup.

pub fn set_backup_progress_in_mega_bytes(self, input: Option<f64>) -> Self

pub fn current_backup_rate_in_mega_bytes_per_second(self, input: f64) -> Self

The number of megabytes per second being transferred to the snapshot backup. Returns 0 for a completed backup.

pub fn set_current_backup_rate_in_mega_bytes_per_second(
self,
input: Option<f64>
) -> Self

pub fn estimated_seconds_to_completion(self, input: i64) -> Self

The estimate of the time remaining before the snapshot backup will complete. Returns 0 for a completed backup.

pub fn set_estimated_seconds_to_completion(self, input: Option<i64>) -> Self

pub fn elapsed_time_in_seconds(self, input: i64) -> Self

The amount of time an in-progress snapshot backup has been running, or the amount of time it took a completed backup to finish.

pub fn set_elapsed_time_in_seconds(self, input: Option<i64>) -> Self

pub fn source_region(self, input: impl Into<String>) -> Self

The source region from which the snapshot was copied.

pub fn set_source_region(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn restorable_node_types(self, input: impl Into<String>) -> Self

pub fn set_restorable_node_types(self, input: Option<Vec<String>>) -> Self

pub fn enhanced_vpc_routing(self, input: bool) -> Self

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

pub fn set_enhanced_vpc_routing(self, input: Option<bool>) -> Self

pub fn maintenance_track_name(self, input: impl Into<String>) -> Self

The name of the maintenance track for the snapshot.

pub fn set_maintenance_track_name(self, input: Option<String>) -> Self

pub fn manual_snapshot_retention_period(self, input: i32) -> Self

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

pub fn set_manual_snapshot_retention_period(self, input: Option<i32>) -> Self

pub fn manual_snapshot_remaining_days(self, input: i32) -> Self

The number of days until a manual snapshot will pass its retention period.

pub fn set_manual_snapshot_remaining_days(self, input: Option<i32>) -> Self

pub fn snapshot_retention_start_time(self, input: Instant) -> Self

A timestamp representing the start of the retention period for the snapshot.

pub fn set_snapshot_retention_start_time(self, input: Option<Instant>) -> Self

pub fn build(self) -> Snapshot

Consumes the builder and constructs a Snapshot