Struct aws_sdk_redshift::model::Cluster

#[non_exhaustive]
pub struct Cluster {

    pub cluster_identifier: Option<String>,
    pub node_type: Option<String>,
    pub cluster_status: Option<String>,
    pub cluster_availability_status: Option<String>,
    pub modify_status: Option<String>,
    pub master_username: Option<String>,
    pub db_name: Option<String>,
    pub endpoint: Option<Endpoint>,
    pub cluster_create_time: Option<Instant>,
    pub automated_snapshot_retention_period: i32,
    pub manual_snapshot_retention_period: i32,
    pub cluster_security_groups: Option<Vec<ClusterSecurityGroupMembership>>,
    pub vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>,
    pub cluster_parameter_groups: Option<Vec<ClusterParameterGroupStatus>>,
    pub cluster_subnet_group_name: Option<String>,
    pub vpc_id: Option<String>,
    pub availability_zone: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub pending_modified_values: Option<PendingModifiedValues>,
    pub cluster_version: Option<String>,
    pub allow_version_upgrade: bool,
    pub number_of_nodes: i32,
    pub publicly_accessible: bool,
    pub encrypted: bool,
    pub restore_status: Option<RestoreStatus>,
    pub data_transfer_progress: Option<DataTransferProgress>,
    pub hsm_status: Option<HsmStatus>,
    pub cluster_snapshot_copy_status: Option<ClusterSnapshotCopyStatus>,
    pub cluster_public_key: Option<String>,
    pub cluster_nodes: Option<Vec<ClusterNode>>,
    pub elastic_ip_status: Option<ElasticIpStatus>,
    pub cluster_revision_number: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub kms_key_id: Option<String>,
    pub enhanced_vpc_routing: bool,
    pub iam_roles: Option<Vec<ClusterIamRole>>,
    pub pending_actions: Option<Vec<String>>,
    pub maintenance_track_name: Option<String>,
    pub elastic_resize_number_of_node_options: Option<String>,
    pub deferred_maintenance_windows: Option<Vec<DeferredMaintenanceWindow>>,
    pub snapshot_schedule_identifier: Option<String>,
    pub snapshot_schedule_state: Option<ScheduleState>,
    pub expected_next_snapshot_schedule_time: Option<Instant>,
    pub expected_next_snapshot_schedule_time_status: Option<String>,
    pub next_maintenance_window_start_time: Option<Instant>,
    pub resize_info: Option<ResizeInfo>,
    pub availability_zone_relocation_status: Option<String>,
    pub cluster_namespace_arn: Option<String>,
    pub total_storage_capacity_in_mega_bytes: Option<i64>,
    pub aqua_configuration: Option<AquaConfiguration>,

}

Describes a cluster.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

cluster_identifier: Option<String>

The unique identifier of the cluster.

node_type: Option<String>

The node type for the nodes in the cluster.

cluster_status: Option<String>

The current state of the cluster. Possible values are the following:

available

available, prep-for-resize

available, resize-cleanup

cancelling-resize

creating

deleting

final-snapshot

hardware-failure

incompatible-hsm

incompatible-network

incompatible-parameters

incompatible-restore

modifying

paused

rebooting

renaming

resizing

rotating-keys

storage-full

updating-hsm

cluster_availability_status: Option<String>

The availability status of the cluster for queries. Possible values are the following:

Available - The cluster is available for queries.

Unavailable - The cluster is not available for queries.

Maintenance - The cluster is intermittently available for queries due to maintenance activities.

Modifying - The cluster is intermittently available for queries due to changes that modify the cluster.

Failed - The cluster failed and is not available for queries.

modify_status: Option<String>

The status of a modify operation, if any, initiated for the cluster.

master_username: Option<String>

The admin user name for the cluster. This name is used to connect to the database that is specified in the DBName parameter.

db_name: Option<String>

The name of the initial database that was created when the cluster was created. This same name is returned for the life of the cluster. If an initial database was not specified, a database named devdev was created by default.

endpoint: Option<Endpoint>

The connection endpoint.

cluster_create_time: Option<Instant>

The date and time that the cluster was created.

automated_snapshot_retention_period: i32

The number of days that automatic cluster snapshots are retained.

manual_snapshot_retention_period: i32

The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

cluster_security_groups: Option<Vec<ClusterSecurityGroupMembership>>

A list of cluster security group that are associated with the cluster. Each security group is represented by an element that contains ClusterSecurityGroup.Name and ClusterSecurityGroup.Status subelements.

Cluster security groups are used when the cluster is not created in an Amazon Virtual Private Cloud (VPC). Clusters that are created in a VPC use VPC security groups, which are listed by the VpcSecurityGroups parameter.

vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>

A list of Amazon Virtual Private Cloud (Amazon VPC) security groups that are associated with the cluster. This parameter is returned only if the cluster is in a VPC.

cluster_parameter_groups: Option<Vec<ClusterParameterGroupStatus>>

The list of cluster parameter groups that are associated with this cluster. Each parameter group in the list is returned with its status.

cluster_subnet_group_name: Option<String>

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

vpc_id: Option<String>

The identifier of the VPC the cluster is in, if the cluster is in a VPC.

availability_zone: Option<String>

The name of the Availability Zone in which the cluster is located.

preferred_maintenance_window: Option<String>

The weekly time range, in Universal Coordinated Time (UTC), during which system maintenance can occur.

pending_modified_values: Option<PendingModifiedValues>

A value that, if present, indicates that changes to the cluster are pending. Specific pending changes are identified by subelements.

cluster_version: Option<String>

The version ID of the Amazon Redshift engine that is running on the cluster.

allow_version_upgrade: bool

A boolean value that, if true, indicates that major version upgrades will be applied automatically to the cluster during the maintenance window.

number_of_nodes: i32

The number of compute nodes in the cluster.

publicly_accessible: bool

A boolean value that, if true, indicates that the cluster can be accessed from a public network.

encrypted: bool

A boolean value that, if true, indicates that data in the cluster is encrypted at rest.

restore_status: Option<RestoreStatus>

A value that describes the status of a cluster restore action. This parameter returns null if the cluster was not created by restoring a snapshot.

data_transfer_progress: Option<DataTransferProgress>

hsm_status: Option<HsmStatus>

A value that reports whether the Amazon Redshift cluster has finished applying any hardware security module (HSM) settings changes specified in a modify cluster command.

Values: active, applying

cluster_snapshot_copy_status: Option<ClusterSnapshotCopyStatus>

A value that returns the destination region and retention period that are configured for cross-region snapshot copy.

cluster_public_key: Option<String>

The public key for the cluster.

cluster_nodes: Option<Vec<ClusterNode>>

The nodes in the cluster.

elastic_ip_status: Option<ElasticIpStatus>

The status of the elastic IP (EIP) address.

cluster_revision_number: Option<String>

The specific revision number of the database in the cluster.

tags: Option<Vec<Tag>>

The list of tags for the cluster.

kms_key_id: Option<String>

The Key Management Service (KMS) key ID of the encryption key used to encrypt data in the cluster.

enhanced_vpc_routing: bool

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

iam_roles: Option<Vec<ClusterIamRole>>

A list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services.

pending_actions: Option<Vec<String>>

Cluster operations that are waiting to be started.

maintenance_track_name: Option<String>

The name of the maintenance track for the cluster.

elastic_resize_number_of_node_options: Option<String>

The number of nodes that you can resize the cluster to with the elastic resize method.

deferred_maintenance_windows: Option<Vec<DeferredMaintenanceWindow>>

Describes a group of DeferredMaintenanceWindow objects.

snapshot_schedule_identifier: Option<String>

A unique identifier for the cluster snapshot schedule.

snapshot_schedule_state: Option<ScheduleState>

The current state of the cluster snapshot schedule.

expected_next_snapshot_schedule_time: Option<Instant>

The date and time when the next snapshot is expected to be taken for clusters with a valid snapshot schedule and backups enabled.

expected_next_snapshot_schedule_time_status: Option<String>

The status of next expected snapshot for clusters having a valid snapshot schedule and backups enabled. Possible values are the following:

OnTrack - The next snapshot is expected to be taken on time.

Pending - The next snapshot is pending to be taken.

next_maintenance_window_start_time: Option<Instant>

The date and time in UTC when system maintenance can begin.

resize_info: Option<ResizeInfo>

Returns the following:

AllowCancelResize: a boolean value indicating if the resize operation can be cancelled.

ResizeType: Returns ClassicResize

availability_zone_relocation_status: Option<String>

Describes the status of the Availability Zone relocation operation.

cluster_namespace_arn: Option<String>

The namespace Amazon Resource Name (ARN) of the cluster.

total_storage_capacity_in_mega_bytes: Option<i64>

The total storage capacity of the cluster in megabytes.

aqua_configuration: Option<AquaConfiguration>

The AQUA (Advanced Query Accelerator) configuration of the cluster.

Implementations

impl Cluster

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Cluster