Struct aws_sdk_redshift::model::Snapshot

#[non_exhaustive]
pub struct Snapshot {

    pub snapshot_identifier: Option<String>,
    pub cluster_identifier: Option<String>,
    pub snapshot_create_time: Option<Instant>,
    pub status: Option<String>,
    pub port: i32,
    pub availability_zone: Option<String>,
    pub cluster_create_time: Option<Instant>,
    pub master_username: Option<String>,
    pub cluster_version: Option<String>,
    pub engine_full_version: Option<String>,
    pub snapshot_type: Option<String>,
    pub node_type: Option<String>,
    pub number_of_nodes: i32,
    pub db_name: Option<String>,
    pub vpc_id: Option<String>,
    pub encrypted: bool,
    pub kms_key_id: Option<String>,
    pub encrypted_with_hsm: bool,
    pub accounts_with_restore_access: Option<Vec<AccountWithRestoreAccess>>,
    pub owner_account: Option<String>,
    pub total_backup_size_in_mega_bytes: f64,
    pub actual_incremental_backup_size_in_mega_bytes: f64,
    pub backup_progress_in_mega_bytes: f64,
    pub current_backup_rate_in_mega_bytes_per_second: f64,
    pub estimated_seconds_to_completion: i64,
    pub elapsed_time_in_seconds: i64,
    pub source_region: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub restorable_node_types: Option<Vec<String>>,
    pub enhanced_vpc_routing: bool,
    pub maintenance_track_name: Option<String>,
    pub manual_snapshot_retention_period: Option<i32>,
    pub manual_snapshot_remaining_days: Option<i32>,
    pub snapshot_retention_start_time: Option<Instant>,

}

Describes a snapshot.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

snapshot_identifier: Option<String>

The snapshot identifier that is provided in the request.

cluster_identifier: Option<String>

The identifier of the cluster for which the snapshot was taken.

snapshot_create_time: Option<Instant>

The time (in UTC format) when Amazon Redshift began the snapshot. A snapshot contains a copy of the cluster data as of this exact time.

status: Option<String>

The snapshot status. The value of the status depends on the API operation used:

CreateClusterSnapshot and CopyClusterSnapshot returns status as "creating".

DescribeClusterSnapshots returns status as "creating", "available", "final snapshot", or "failed".

DeleteClusterSnapshot returns status as "deleted".

port: i32

The port that the cluster is listening on.

availability_zone: Option<String>

The Availability Zone in which the cluster was created.

cluster_create_time: Option<Instant>

The time (UTC) when the cluster was originally created.

master_username: Option<String>

The admin user name for the cluster.

cluster_version: Option<String>

The version ID of the Amazon Redshift engine that is running on the cluster.

engine_full_version: Option<String>

The cluster version of the cluster used to create the snapshot. For example, 1.0.15503.

snapshot_type: Option<String>

The snapshot type. Snapshots created using CreateClusterSnapshot and CopyClusterSnapshot are of type "manual".

node_type: Option<String>

The node type of the nodes in the cluster.

number_of_nodes: i32

The number of nodes in the cluster.

db_name: Option<String>

The name of the database that was created when the cluster was created.

vpc_id: Option<String>

The VPC identifier of the cluster if the snapshot is from a cluster in a VPC. Otherwise, this field is not in the output.

encrypted: bool

If true, the data in the snapshot is encrypted at rest.

kms_key_id: Option<String>

The Key Management Service (KMS) key ID of the encryption key that was used to encrypt data in the cluster from which the snapshot was taken.

encrypted_with_hsm: bool

A boolean that indicates whether the snapshot data is encrypted using the HSM keys of the source cluster. true indicates that the data is encrypted using HSM keys.

accounts_with_restore_access: Option<Vec<AccountWithRestoreAccess>>

A list of the Amazon Web Services accounts authorized to restore the snapshot. Returns null if no accounts are authorized. Visible only to the snapshot owner.

owner_account: Option<String>

For manual snapshots, the Amazon Web Services account used to create or copy the snapshot. For automatic snapshots, the owner of the cluster. The owner can perform all snapshot actions, such as sharing a manual snapshot.

total_backup_size_in_mega_bytes: f64

The size of the complete set of backup data that would be used to restore the cluster.

actual_incremental_backup_size_in_mega_bytes: f64

The size of the incremental backup.

backup_progress_in_mega_bytes: f64

The number of megabytes that have been transferred to the snapshot backup.

current_backup_rate_in_mega_bytes_per_second: f64

The number of megabytes per second being transferred to the snapshot backup. Returns 0 for a completed backup.

estimated_seconds_to_completion: i64

The estimate of the time remaining before the snapshot backup will complete. Returns 0 for a completed backup.

elapsed_time_in_seconds: i64

The amount of time an in-progress snapshot backup has been running, or the amount of time it took a completed backup to finish.

source_region: Option<String>

The source region from which the snapshot was copied.

tags: Option<Vec<Tag>>

The list of tags for the cluster snapshot.

restorable_node_types: Option<Vec<String>>

The list of node types that this cluster snapshot is able to restore into.

enhanced_vpc_routing: bool

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

maintenance_track_name: Option<String>

The name of the maintenance track for the snapshot.

manual_snapshot_retention_period: Option<i32>

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

manual_snapshot_remaining_days: Option<i32>

The number of days until a manual snapshot will pass its retention period.

snapshot_retention_start_time: Option<Instant>

A timestamp representing the start of the retention period for the snapshot.

Implementations

impl Snapshot

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Snapshot