Module aws_sdk_redshift::operation

Structs

AcceptReservedNodeExchange

Exchanges a DC1 Reserved Node for a DC2 Reserved Node with no changes to the

AddPartner

Adds a partner integration to a cluster.

AssociateDataShareConsumer

From a datashare consumer account, associates a datashare with the

AuthorizeClusterSecurityGroupIngress

Adds an inbound (ingress) rule to an Amazon Redshift security group. Depending on whether

AuthorizeDataShare

From a data producer account, authorizes the sharing of a datashare with one or more

AuthorizeEndpointAccess

Grants access to a cluster.

AuthorizeSnapshotAccess

Authorizes the specified Amazon Web Services account to restore the specified

BatchDeleteClusterSnapshots

Deletes a set of cluster snapshots.

BatchModifyClusterSnapshots

Modifies the settings for a set of cluster snapshots.

CancelResize

Cancels a resize operation for a cluster.

CopyClusterSnapshot

Copies the specified automated cluster snapshot to a new manual cluster snapshot.

CreateAuthenticationProfile

Creates an authentication profile with the specified parameters.

CreateCluster

Creates a new cluster with the specified parameters.

CreateClusterParameterGroup

Creates an Amazon Redshift parameter group.

CreateClusterSecurityGroup

Creates a new Amazon Redshift security group. You use security groups to control access

CreateClusterSnapshot

Creates a manual snapshot of the specified cluster. The cluster must be in the

CreateClusterSubnetGroup

Creates a new Amazon Redshift subnet group. You must provide a list of one or more

CreateEndpointAccess

Creates a Redshift-managed VPC endpoint.

CreateEventSubscription

Creates an Amazon Redshift event notification subscription. This action requires an ARN

CreateHsmClientCertificate

Creates an HSM client certificate that an Amazon Redshift cluster will use to connect to

CreateHsmConfiguration

Creates an HSM configuration that contains the information required by an Amazon Redshift

CreateScheduledAction

Creates a scheduled action. A scheduled action contains a schedule and an Amazon Redshift API action.

CreateSnapshotCopyGrant

Creates a snapshot copy grant that permits Amazon Redshift to use a customer master key

CreateSnapshotSchedule

Create a snapshot schedule that can be associated to a cluster and which overrides the default system backup schedule.

CreateTags

Adds tags to a cluster.

CreateUsageLimit

Creates a usage limit for a specified Amazon Redshift feature on a cluster.

DeauthorizeDataShare

From the producer account, removes authorization from the specified datashare.

DeleteAuthenticationProfile

Deletes an authentication profile.

DeleteCluster

Deletes a previously provisioned cluster without its final snapshot being created. A successful response from the web

DeleteClusterParameterGroup

Deletes a specified Amazon Redshift parameter group.

DeleteClusterSecurityGroup

Deletes an Amazon Redshift security group.

DeleteClusterSnapshot

Deletes the specified manual snapshot. The snapshot must be in the

DeleteClusterSubnetGroup

Deletes the specified cluster subnet group.

DeleteEndpointAccess

Deletes a Redshift-managed VPC endpoint.

DeleteEventSubscription

Deletes an Amazon Redshift event notification subscription.

DeleteHsmClientCertificate

Deletes the specified HSM client certificate.

DeleteHsmConfiguration

Deletes the specified Amazon Redshift HSM configuration.

DeletePartner

Deletes a partner integration from a cluster. Data can still flow to the cluster until the integration is deleted at the partner's website.

DeleteScheduledAction

Deletes a scheduled action.

DeleteSnapshotCopyGrant

Deletes the specified snapshot copy grant.

DeleteSnapshotSchedule

Deletes a snapshot schedule.

DeleteTags

Deletes tags from a resource. You must provide the ARN of the resource

DeleteUsageLimit

Deletes a usage limit from a cluster.

DescribeAccountAttributes

Returns a list of attributes attached to an account

DescribeAuthenticationProfiles

Describes an authentication profile.

DescribeClusterDbRevisions

Returns an array of ClusterDbRevision objects.

DescribeClusterParameterGroups

Returns a list of Amazon Redshift parameter groups, including parameter groups you

DescribeClusterParameters

Returns a detailed list of parameters contained within the specified Amazon Redshift

DescribeClusterSecurityGroups

Returns information about Amazon Redshift security groups. If the name of a security

DescribeClusterSnapshots

Returns one or more snapshot objects, which contain metadata about your cluster

DescribeClusterSubnetGroups

Returns one or more cluster subnet group objects, which contain metadata about your

DescribeClusterTracks

Returns a list of all the available maintenance tracks.

DescribeClusterVersions

Returns descriptions of the available Amazon Redshift cluster versions. You can call this

DescribeClusters

Returns properties of provisioned clusters including general cluster properties,

DescribeDataShares

Shows the status of any inbound or outbound datashares available in the specified

DescribeDataSharesForConsumer

Returns a list of datashares where the account identifier being called is a consumer account identifier.

DescribeDataSharesForProducer

Returns a list of datashares when the account identifier being called is a producer account identifier.

DescribeDefaultClusterParameters

Returns a list of parameter settings for the specified parameter group

DescribeEndpointAccess

Describes a Redshift-managed VPC endpoint.

DescribeEndpointAuthorization

Describes an endpoint authorization.

DescribeEventCategories

Displays a list of event categories for all event source types, or for a specified

DescribeEventSubscriptions

Lists descriptions of all the Amazon Redshift event notification subscriptions for a

DescribeEvents

Returns events related to clusters, security groups, snapshots, and parameter

DescribeHsmClientCertificates

Returns information about the specified HSM client certificate. If no certificate

DescribeHsmConfigurations

Returns information about the specified Amazon Redshift HSM configuration. If no

DescribeLoggingStatus

Describes whether information, such as queries and connection attempts, is being

DescribeNodeConfigurationOptions

Returns properties of possible node configurations such as node type, number of nodes, and

DescribeOrderableClusterOptions

Returns a list of orderable cluster options. Before you create a new cluster you

DescribePartners

Returns information about the partner integrations defined for a cluster.

DescribeReservedNodeOfferings

Returns a list of the available reserved node offerings by Amazon Redshift with their

DescribeReservedNodes

Returns the descriptions of the reserved nodes.

DescribeResize

Returns information about the last resize operation for the specified cluster. If

DescribeScheduledActions

Describes properties of scheduled actions.

DescribeSnapshotCopyGrants

Returns a list of snapshot copy grants owned by the Amazon Web Services account in the destination

DescribeSnapshotSchedules

Returns a list of snapshot schedules.

DescribeStorage

Returns account level backups storage size and provisional storage.

DescribeTableRestoreStatus

Lists the status of one or more table restore requests made using the RestoreTableFromClusterSnapshot API action. If you don't specify a value

DescribeTags

Returns a list of tags. You can return tags from a specific resource by specifying

DescribeUsageLimits

Shows usage limits on a cluster.

DisableLogging

Stops logging information, such as queries and connection attempts, for the

DisableSnapshotCopy

Disables the automatic copying of snapshots from one region to another region for a

DisassociateDataShareConsumer

From a consumer account, remove association for the specified datashare.

EnableLogging

Starts logging information, such as queries and connection attempts, for the

EnableSnapshotCopy

Enables the automatic copy of snapshots from one region to another region for a

GetClusterCredentials

Returns a database user name and temporary password with temporary authorization to

GetReservedNodeExchangeOfferings

Returns an array of DC2 ReservedNodeOfferings that matches the payment type, term,

ModifyAquaConfiguration

Modifies whether a cluster can use AQUA (Advanced Query Accelerator).

ModifyAuthenticationProfile

Modifies an authentication profile.

ModifyCluster

Modifies the settings for a cluster.

ModifyClusterDbRevision

Modifies the database revision of a cluster. The database revision is a unique

ModifyClusterIamRoles

Modifies the list of Identity and Access Management (IAM) roles that can be

ModifyClusterMaintenance

Modifies the maintenance settings of a cluster.

ModifyClusterParameterGroup

Modifies the parameters of a parameter group. For the parameters parameter, it can't contain ASCII characters.

ModifyClusterSnapshot

Modifies the settings for a snapshot.

ModifyClusterSnapshotSchedule

Modifies a snapshot schedule for a cluster.

ModifyClusterSubnetGroup

Modifies a cluster subnet group to include the specified list of VPC subnets. The

ModifyEndpointAccess

Modifies a Redshift-managed VPC endpoint.

ModifyEventSubscription

Modifies an existing Amazon Redshift event notification subscription.

ModifyScheduledAction

Modifies a scheduled action.

ModifySnapshotCopyRetentionPeriod

Modifies the number of days to retain snapshots in the destination Amazon Web Services Region after

ModifySnapshotSchedule

Modifies a snapshot schedule. Any schedule associated with a cluster is modified

ModifyUsageLimit

Modifies a usage limit in a cluster.

PauseCluster

Pauses a cluster.

PurchaseReservedNodeOffering

Allows you to purchase reserved nodes. Amazon Redshift offers a predefined set of

RebootCluster

Reboots a cluster. This action is taken as soon as possible. It results in a

RejectDataShare

From the consumer account, rejects the specified datashare.

ResetClusterParameterGroup

Sets one or more parameters of the specified parameter group to their default

ResizeCluster

Changes the size of the cluster. You can change the cluster's type, or change the

RestoreFromClusterSnapshot

Creates a new cluster from a snapshot. By default, Amazon Redshift creates the resulting

RestoreTableFromClusterSnapshot

Creates a new table from a table in an Amazon Redshift cluster snapshot. You must

ResumeCluster

Resumes a paused cluster.

RevokeClusterSecurityGroupIngress

Revokes an ingress rule in an Amazon Redshift security group for a previously authorized

RevokeEndpointAccess

Revokes access to a cluster.

RevokeSnapshotAccess

Removes the ability of the specified Amazon Web Services account to restore the specified

RotateEncryptionKey

Rotates the encryption keys for a cluster.

UpdatePartnerStatus

Updates the status of a partner integration.